Transforms a fixest_multi object into a list

Extracts the results from a fixest_multi object and place them into a list.

# S3 method for fixest_multi
as.list(x, ...)

Arguments

x
A fixest_multi object, obtained from a fixest estimation leading to multiple results.

...
Not currently used.

Value

Returns a list containing all the results of the multiple estimations.

See also

Examples

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res = feols(y ~ csw(x1, x2, x3), base, split = ~species)

# All the results at once
as.list(res)

#> [[1]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value Pr(>|t|))    
#> (Intercept)  2.63900   0.310014  8.5125  3.74e-11 ***
#> x1           0.69049   0.089899  7.6807  6.71e-10 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.233723   Adj. R2: 0.542029
#> 
#> [[2]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value  Pr(>|t|))    
#> (Intercept) 2.303700   0.385294  5.9792 2.8900e-07 ***
#> x1          0.667416   0.090356  7.3865 2.1300e-09 ***
#> x2          0.283419   0.197224  1.4370 1.5733e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.228751   Adj. R2: 0.551971
#> 
#> [[3]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error  t value   Pr(>|t|))    
#> (Intercept) 2.351900   0.392868 5.986500 3.03000e-07 ***
#> x1          0.654835   0.092447 7.083300 6.83000e-09 ***
#> x2          0.237560   0.208019 1.142000 2.59359e-01    
#> x3          0.252126   0.346864 0.726873 4.70987e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.227449   Adj. R2: 0.547429
#> 
#> [[4]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value Pr(>|t|))    
#> (Intercept) 3.539700   0.562874  6.2887  9.07e-08 ***
#> x1          0.865078   0.201938  4.2839  8.80e-05 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.434612   Adj. R2: 0.261511
#> 
#> [[5]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value   Pr(>|t|))    
#> (Intercept) 2.116400   0.494256  4.2821 0.000091000 ***
#> x1          0.247642   0.186839  1.3254 0.191435000    
#> x2          0.735587   0.124768  5.8956 0.000000387 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.329521   Adj. R2: 0.566438
#> 
#> [[6]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>              Estimate Std. Error t value  Pr(>|t|))    
#> (Intercept)  1.895500   0.507055  3.7383 0.00051100 ***
#> x1           0.386858   0.204545  1.8913 0.06489000 .  
#> x2           0.908337   0.165432  5.4907 0.00000167 ***
#> x3          -0.679224   0.435382 -1.5601 0.12559900    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.321135   Adj. R2: 0.579273
#> 
#> [[7]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value Pr(>|t|))    
#> (Intercept) 3.906800   0.757061  5.1605  4.66e-06 ***
#> x1          0.901534   0.253106  3.5619  8.43e-04 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.559836   Adj. R2: 0.192579
#> 
#> [[8]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value   Pr(>|t|))    
#> (Intercept) 0.624782   0.524867  1.1904 2.39882e-01    
#> x1          0.259954   0.153338  1.6953 9.66340e-02 .  
#> x2          0.934819   0.089602 10.4330 8.01000e-14 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.307439   Adj. R2: 0.75132
#> 
#> [[9]]
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>              Estimate Std. Error   t value   Pr(>|t|))    
#> (Intercept)  0.699883   0.533601  1.311600 1.96156e-01    
#> x1           0.330337   0.174329  1.894900 6.44000e-02 .  
#> x2           0.945536   0.090722 10.422000 1.07000e-13 ***
#> x3          -0.169753   0.198072 -0.857023 3.95875e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.305014   Adj. R2: 0.749908
#>