Sample data for staggered difference in difference

This data has been generated to illustrate the Sun and Abraham (Journal of Econometrics, 2021) method for staggered difference-in-difference. This is a balanced panel of 95 individuals and 10 periods. Half the individuals are treated. For those treated, the treatment date can vary from the second to the last period. The effect of the treatment depends on the time since the treatment: it is first negative and then increasing.

Usage

data(base_stagg, package = "fixest")

Format

base_stagg is a data frame with 950 observations and 7 variables:

id: panel identifier.

year: from 1 to 10.

year_treated: the period at which the individual is treated.

time_to_treatment: different between the year and the treatment year.

treated: indicator taking value 1 if the individual is treated, 0 otherwise.

treatment_effect_true: true effect of the treatment.

x1: explanatory variable, correlated with the period.

y: the dependent variable affected by the treatment.

Source

This data has been generated from R.