Extracts the coefficients of fixest_multi objects

Utility to extract the coefficients of multiple estimations and rearrange them into a matrix.

# S3 method for fixest_multi
coef(object, keep, drop, order, ...)

# S3 method for fixest_multi
coefficients(object, keep, drop, order, ...)

Arguments

object
A fixest_multi object. Obtained from a multiple estimation.

keep
Character vector. This element is used to display only a subset of variables. This should be a vector of regular expressions (see regex help for more info). Each variable satisfying any of the regular expressions will be kept. This argument is applied post aliasing (see argument dict). Example: you have the variable x1 to x55 and want to display only x1 to x9, then you could use keep = "x[[:digit:]]$". If the first character is an exclamation mark, the effect is reversed (e.g. keep = "!Intercept" means: every variable that does not contain “Intercept” is kept). See details.

drop
Character vector. This element is used if some variables are not to be displayed. This should be a vector of regular expressions (see regex help for more info). Each variable satisfying any of the regular expressions will be discarded. This argument is applied post aliasing (see argument dict). Example: you have the variable x1 to x55 and want to display only x1 to x9, then you could use drop = "x[[:digit:]]{2}". If the first character is an exclamation mark, the effect is reversed (e.g. drop = "!Intercept" means: every variable that does not contain “Intercept” is dropped). See details.

order
Character vector. This element is used if the user wants the variables to be ordered in a certain way. This should be a vector of regular expressions (see regex help for more info). The variables satisfying the first regular expression will be placed first, then the order follows the sequence of regular expressions. This argument is applied post aliasing (see argument dict). Example: you have the following variables: month1 to month6, then x1 to x5, then year1 to year6. If you want to display first the x's, then the years, then the months you could use: order = c("x", "year"). If the first character is an exclamation mark, the effect is reversed (e.g. order = "!Intercept" means: every variable that does not contain “Intercept” goes first). See details.

...
Not currently used.

Examples

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")

# A multiple estimation
est = feols(y ~ x1 + csw0(x2, x3), base)

# Getting all the coefficients at once,
# each row is a model
coef(est)

#>      (Intercept)         x1       x2         x3
#> [1,]    6.526223 -0.2233611       NA         NA
#> [2,]    2.249140  0.5955247 0.471920         NA
#> [3,]    1.855997  0.6508372 0.709132 -0.5564827

# Example of keep/drop/order
coef(est, keep = "Int|x1", order = "x1")

#>              x1 (Intercept)
#> [1,] -0.2233611    6.526223
#> [2,]  0.5955247    2.249140
#> [3,]  0.6508372    1.855997

# To change the order of the model, use fixest_multi
# extraction tools:
coef(est[rhs = .N:1])

#>      (Intercept)         x1       x2         x3
#> [1,]    1.855997  0.6508372 0.709132 -0.5564827
#> [2,]    2.249140  0.5955247 0.471920         NA
#> [3,]    6.526223 -0.2233611       NA         NA