Obtain various statistics from an estimation

Set of functions to directly extract some commonly used statistics, like the p-value or the table of coefficients, from estimations. This was first implemented for fixest estimations, but has some support for other models.

coeftable(object, se, cluster, ...)

ctable(object, se, cluster, ...)

pvalue(object, se, cluster, ...)

tstat(object, se, cluster, ...)

se(object, se, cluster, ...)

Arguments

object
An estimation. For example obtained from feols.

se
[Fixest specific.] Character scalar. Which kind of standard error should be computed: “standard”, “White”, “cluster”, “twoway”, “threeway” or “fourway”? By default if there are clusters in the estimation: se = "cluster", otherwise se = "standard". Note that this argument can be implicitly deduced from the argument cluster.

cluster
[Fixest specific.] Tells how to cluster the standard-errors (if clustering is requested). Can be either a list of vectors, a character vector of variable names, a formula or an integer vector. Assume we want to perform 2-way clustering over var1 and var2 contained in the data.frame base used for the estimation. All the following cluster arguments are valid and do the same thing: cluster = base[, c("var1, "var2")]}, \code{cluster = c("var1, "var2"), cluster = ~var1+var2. If the two variables were used as clusters in the estimation, you could further use cluster = 1:2 or leave it blank with se = "twoway" (assuming var1 [resp. var2] was the 1st [res. 2nd] cluster).

...
Other arguments to be passed to summary.

Value

Returns a table of coefficients, with in rows the variables and four columns: the estimate, the standard-error, the t-statistic and the p-value.

Details

This set of functions is primarily constructed for fixest estimations. Although it can work for non-fixest estimations, support for exotic estimation procedures that do not report standardized coefficient tables is highly limited.

Functions

pvalue: Extracts the p-value of an estimation

tstat: Extracts the t-statistics of an estimation

se: Extracts the standard-error of an estimation

Examples

# Some data and estimation
data(trade)
est = fepois(Euros ~ log(dist_km) | Origin^Product + Year, trade)

#
# Coeftable/se/tstat/pvalue
#

# Default is clustering along Origin^Product
coeftable(est)

#>               Estimate Std. Error   z value      Pr(>|z|)
#> log(dist_km) -1.023957 0.04728994 -21.65275 2.416303e-103

se(est)

#> log(dist_km) 
#>   0.04728994 

tstat(est)

#> log(dist_km) 
#>    -21.65275 

pvalue(est)

#>  log(dist_km) 
#> 2.416303e-103 

# Now with two-way clustered standard-errors
#  and using ctable(), the alias to coeftable()

ctable(est, cluster = ~Origin + Product)

#>               Estimate Std. Error   z value     Pr(>|z|)
#> log(dist_km) -1.023957 0.09128554 -11.21708 3.735951e-29

se(est, cluster = ~Origin + Product)

#> log(dist_km) 
#>   0.09128554 

pvalue(est, cluster = ~Origin + Product)

#> log(dist_km) 
#> 3.735951e-29 

tstat(est, cluster = ~Origin + Product)

#> log(dist_km) 
#>    -11.21708 

# Or you can cluster only once:
est_sum = summary(est, cluster = ~Origin + Product)
ctable(est_sum)

#>               Estimate Std. Error   z value     Pr(>|z|)
#> log(dist_km) -1.023957 0.09128554 -11.21708 3.735951e-29

se(est_sum)

#> log(dist_km) 
#>   0.09128554 

tstat(est_sum)

#> log(dist_km) 
#>    -11.21708 

pvalue(est_sum)

#> log(dist_km) 
#> 3.735951e-29