Simple and powerful string manipulation with the dot square bracket operator

Simple and powerful string manipulation with the dot square bracket operator

dsb(..., frame = parent.frame(), sep = "", vectorize = FALSE, nest = TRUE)

Arguments

...
Character scalars that will be collapsed with the argument sep. You can use ".[x]" within each character string to insert the value of x in the string. You can add string operations in each ".[]" instance with the syntax "'arg'op ? x" (resp. "'arg'op ! x") to apply the operation 'op' with the argument 'arg' to x (resp. the verbatim of x). Otherwise, what to say? Ah, nesting is enabled, and since there's over 30 operators, it's bit complicated to sort you out in this small space. But type dsb("--help") to prompt an (almost) extensive help.

frame
An environment used to evaluate the variables in ".[]".

sep
Character scalar, default is "". It is used to collapse all the elements in ....

vectorize
Logical, default is FALSE. If TRUE, Further, elements in ... are NOT collapsed together, but instead vectorised.

nest
Logical, default is TRUE. Whether the original character strings should be nested into a ".[]". If TRUE, then things like dsb("S!one, two") are equivalent to dsb(".[S!one, two]") and hence create the vector c("one", "two").

Value

It returns a character vector whose length depends on the elements and operations in ".[]".

Examples