Extracts the scores from a fixest estimation

Extracts the scores from a fixest estimation.

Usage

# S3 method for class 'fixest'
estfun(x, ...)

Arguments

x

A fixest object, obtained for instance from feols.

...

Not currently used.

Value

Returns a matrix of the same number of rows as the number of observations used for the estimation, and the same number of columns as there were variables.

Examples

data(iris)
est = feols(Petal.Length ~ Petal.Width + Sepal.Width, iris)
head(estfun(est))
#>      (Intercept)  Petal.Width Sepal.Width
#> [1,] -0.04666458 -0.009332917 -0.16332605
#> [2,] -0.22418187 -0.044836375 -0.67254562
#> [3,] -0.25317496 -0.050634992 -0.81015987
#> [4,] -0.08867842 -0.017735683 -0.27490309
#> [5,] -0.01116113 -0.002232225 -0.04018005
#> [6,] -0.03577286 -0.014309142 -0.13951414