Formatted dimension

Prints the dimension of a data set, in an user-readable way

Usage

fdim(x)

Arguments

x

An R object, usually a data.frame (but can also be a vector).

Value

It does not return anything, the output is directly printed on the console.

Author

Laurent Berge

Examples

fdim(iris)
#> 150 rows and 5 columns

fdim(iris$Species)
#> 150 obs