Fixed-effects GLM estimations

Estimates GLM models with any number of fixed-effects.

feglm(fml, data, family = "poisson", offset, weights, panel.id,
  start = NULL, etastart = NULL, mustart = NULL, fixef,
  fixef.tol = 1e-06, fixef.iter = 1000, glm.iter = 25,
  glm.tol = 1e-08, na_inf.rm = getFixest_na_inf.rm(),
  nthreads = getFixest_nthreads(), warn = TRUE,
  notes = getFixest_notes(), verbose = 0, combine.quick, ...)

feglm.fit(y, X, fixef_mat, family = "poisson", offset, weights,
  start = NULL, etastart = NULL, mustart = NULL, fixef.tol = 1e-06,
  fixef.iter = 1000, glm.iter = 25, glm.tol = 1e-08,
  na_inf.rm = getFixest_na_inf.rm(), nthreads = getFixest_nthreads(),
  warn = TRUE, notes = getFixest_notes(), verbose = 0, ...)

fepois(fml, data, offset, weights, panel.id, start = NULL,
  etastart = NULL, mustart = NULL, fixef, fixef.tol = 1e-06,
  fixef.iter = 1000, glm.iter = 25, glm.tol = 1e-08,
  na_inf.rm = getFixest_na_inf.rm(), nthreads = getFixest_nthreads(),
  warn = TRUE, notes = getFixest_notes(), verbose = 0, combine.quick,
  ...)

Arguments

fml
A formula representing the relation to be estimated. For example: fml = z~x+y. To include fixed-effects, insert them in this formula using a pipe: e.g. fml = z~x+y | fe_1+fe_2. You can combine two clusters with ^: e.g. fml = z~x+y|fe_1^fe_2, see details. You can also use variables with varying slopes using square brackets: e.g. in fml = z~y|fe_1[x] + fe_2 the variable x will have one coefficient for each value of fe_1 -- if you use varying slopes, please have a look at the details section (can't describe it all here).

data
A data.frame containing the necessary variables to run the model. The variables of the non-linear right hand side of the formula are identified with this data.frame names. Can also be a matrix.

family
Family to be used for the estimation. Defaults to poisson(). See family for details of family functions.

offset
A formula or a numeric vector. An offset can be added to the estimation. If equal to a formula, it should be of the form (for example) ~0.5*x**2. This offset is linearly added to the elements of the main formula 'fml'.

weights
A formula or a numeric vector. Each observation can be weighted, the weights must be greater than 0. If equal to a formula, it should be of one-sided: for example ~ var_weight.

panel.id
The panel identifiers. Can either be: i) a one sided formula (e.g. panel.id~id+time), ii) a character vector of length 2 (e.g. panel.id=c('id', 'time'), or iii) a character scalar of two variables separated by a comma (e.g. panel.id='id,time'). Note that you can combine variables with ^ only inside formulas (see the dedicated section in feols).

start
Starting values for the coefficients. Can be: i) a numeric of length 1 (e.g. start = 0), ii) a numeric vector of the exact same length as the number of variables, or iii) a named vector of any length (the names will be used to initialize the appropriate coefficients). Default is missing.

etastart
Numeric vector of the same length as the data. Starting values for the linear predictor. Default is missing.

mustart
Numeric vector of the same length as the data. Starting values for the vector of means. Default is missing.

fixef
Character vector. The name/s of a/some variable/s within the dataset to be used as fixed-effects. These variables should contain the identifier of each observation (e.g., think of it as a panel identifier).

fixef.tol
Precision used to obtain the fixed-effects (ie cluster coefficients). Defaults to 1e-5. It corresponds to the maximum absolute difference allowed between two coefficients of successive iterations. Argument fixef.tol cannot be lower than 10000*.Machine$double.eps. Note that this parameter is dynamically controlled by the algorithm.

fixef.iter
Maximum number of iterations in the step obtaining the fixed-effects (only in use for 2+ clusters). Default is 10000.

glm.iter
Number of iterations of the glm algorithm. Default is 25.

glm.tol
Tolerance level for the glm algorithm. Default is 1e-8.

na_inf.rm
Logical, default is TRUE. If the variables necessary for the estimation contain NA/Infs and na_inf.rm = TRUE, then all observations containing NA are removed prior to estimation and a note is displayed detailing the number of observations removed. Otherwise, an error is raised.

nthreads
Integer: Number of nthreads to be used (accelerates the algorithm via the use of openMP routines). The default is to use the total number of nthreads available minus two. You can set permanently the number of nthreads used within this package using the function setFixest_nthreads.

warn
Logical, default is TRUE. Whether warnings should be displayed (concerns warnings relating to: convergence state, collinearity issues and observation removal due to only 0/1 outcomes or presence of NA values).

notes
Logical. By default, two notes are displayed: when NAs are removed (to show additional information) and when some observations are removed because of only 0 (or 0/1) outcomes in a fixed-effect (in Poisson/Neg. Bin./Logit models). To avoid displaying these messages, you can set notes = FALSE. You can remove these messages permanently by using setFixest_notes(FALSE).

verbose
Integer, default is 0. It represents the level of information that should be reported during the optimisation process. If verbose=0: nothing is reported. If verbose=1: the value of the coefficients and the likelihood are reported. If verbose=2: 1 + information on the computing time of the null model, the cluster coefficients and the hessian are reported.

combine.quick
Logical. When you combine different variables to transform them into a single fixed-effects you can do e.g. y ~ x | paste(var1, var2). The algorithm provides a shorthand to do the same operation: y ~ x | var1^var2. Because pasting variables is a costly operation, the internal algorithm may use a numerical trick to hasten the process. The cost of doing so is that you lose the labels. If you are interested in getting the value of the fixed-effects coefficients after the estimation, you should use combine.quick = FALSE. By default it is equal to FALSE if the number of observations is lower than 50,000, and to TRUE otherwise.

...
Not currently used.

y
Numeric vector of the dependent variable.

X
Numeric matrix of the regressors.

fixef_mat
Matrix/data.frame of the fixed-effects.

Details

The core of the GLM are the weighted OLS estimations. These estimations are performed with feols. The method used to demean each variable along the fixed-effects is based on Berge (2018), since this is the same problem to solve as for the Gaussian case in a ML setup.

Functions

feglm.fit: Matrix method for fixed-effects GLM estimation

fepois: Fixed-effects Poisson estimation

Combining the fixed-effects

You can combine two variables to make it a new fixed-effect using ^. The syntax is as follows: fe_1^fe_2. Here you created a new variable which is the combination of the two variables fe_1 and fe_2. This is identical to doing paste0(fe_1, "_", fe_2) but more convenient.

Note that pasting is a costly operation, especially for large data sets. Thus, the internal algorithm uses a numerical trick which is fast, but the drawback is that the identity of each observation is lost (i.e. they are now equal to a meaningless number instead of being equal to paste0(fe_1, "_", fe_2)). These “identities” are useful only if you're interested in the value of the fixed-effects (that you can extract with fixef.fixest). If you're only interested in coefficients of the variables, it doesn't matter. Anyway, you can use combine.quick = FALSE to tell the internal algorithm to use paste instead of the numerical trick. By default, the numerical trick is performed only for large data sets.

Varying slopes

You can add variables with varying slopes in the fixed-effect part of the formula. The syntax is as follows: cluster_var[var1, var2]. Here the variables var1 and var2 will be with varying slopes (one slope per value in cluster_var) and the fixed-effect cluster_var will also be added.

To add only the variables with varying slopes and not the fixed-effect, use double square brackets: cluster_var[[var1, var2]].

In other words:

cluster_var[var1, var2] is equivalent to cluster_var + cluster_var[[var1]] + cluster_var[[var2]]

cluster_var[[var1, var2]] is equivalent to cluster_var[[var1]] + cluster_var[[var2]]

Lagging variables

To use leads/lags of variables in the estimation, you can: i) either provide the argument panel.id, ii) either set you data set as a panel with the function panel. Doing either of the two will give you acceess to the lagging functions l and f.

You can provide several leads/lags at once: e.g. if your formula is equal to f(y) ~ l(x, -1:1), it means that the dependent variable is equal to the lead of y, and you will have as explanatory variables the lead of x1, x1 and the lag of x1. See the examples in function l for more details.

Interactions

You can interact a variable with a "factor-like" variable by using the syntax var::fe(ref), where fe is the variable to be interacted with and the argument ref is a value of fe taken as a reference.

The full syntax is: var::fe(ref, confirm). You have two arguments, ref and confirm, the two are optional. The argument confirm is there to avoid mistakenly estimating a model with (too) many variables. If the variable fe takes over 100 different values, then you have to add the argument confirm = TRUE. It is important to note that *if you do not care about the standard-errors of the interactions*, then you can add interactions in the fixed-effects part of the formula (using the syntax fe[[var]], as explained in the section “Varying slopes”).

Introducing interactions with this syntax leads to a different display of the interacted values in etable and offers a special representation of the interacted coefficients in the function coefplot. See examples.

The syntax var::fe(ref) is in fact a shorthand for interact(var, fe, ref), you have more information in interact help pages.

References

Berge, Laurent, 2018, "Efficient estimation of maximum likelihood models with multiple fixed-effects: the R package FENmlm." CREA Discussion Papers, 13 (https://wwwen.uni.lu/content/download/110162/1299525/file/2018_13).

For models with multiple fixed-effects:

Gaure, Simen, 2013, "OLS with multiple high dimensional category variables", Computational Statistics & Data Analysis 66 pp. 8--18

See also

Examples

# Default is a poisson model
res = feglm(Sepal.Length ~ Sepal.Width + Petal.Length | Species, iris)

# You could also use fepois
res_pois = fepois(Sepal.Length ~ Sepal.Width + Petal.Length | Species, iris)

# With the fit method:
res_fit = feglm.fit(iris$Sepal.Length, iris[, 2:3], iris$Species)

# All results are identical:
etable(res, res_pois, res_fit)

#>                                   res          res_pois           res_fit
#> Dependent Var.:          Sepal.Length      Sepal.Length iris$Sepal.Length
#>                                                                          
#> Sepal.Width          0.0778* (0.0338)  0.0778* (0.0338)  0.0778* (0.0338)
#> Petal.Length        0.1221*** (0.015) 0.1221*** (0.015) 0.1221*** (0.015)
#> Fixed-Effects:      ----------------- ----------------- -----------------
#> Species                           Yes               Yes                No
#> iris$Species                       No                No               Yes
#> ___________________ _________________ _________________ _________________
#> Observations                      150               150               150
#> S.E. type              1-way: Species    1-way: Species 1-way: iris$Spe..
#> Squared Corr.                   0.863             0.863             0.863
#> Pseudo R2                     0.02676           0.02676           0.02676
#> BIC                            595.81            595.81            595.81