Functions exported from nlme to implement fixest methods

The package fixest uses the fixef method from nlme. Unfortunately, re-exporting this method is required in order not to attach package nlme.

Details

Here is the help from package nlme: fixef. The help from package fixest is here: fixef.fixest.

Note

I could find this workaround thanks to the package plm.