Interact variables with factors

Interacts a variable with another treated as a factor, and sets a reference

i(var, fe, ref, confirm = FALSE)

interact(var, fe, ref, confirm = FALSE)

Arguments

var
A vector.

fe
A vector (of any type). Must be of the same length as var.

ref
A single value that belongs to the interacted variable (fe). Can be missing.

confirm
Logical, default is FALSE. If the factor variable has over 100 cases, you need to use confirm = TRUE to carry on.

Value

It returns a matrix with number of rows the length of var. The number of columns is equal to the number of cases contained in fe minus the reference.

Shorthand in fixest estimations

In fixest estimations, instead of using i(var, fe, ref), you can instead use the following writing var::fe(ref).

See also

Examples

#
# Simple illustration
#

x = rnorm(10)
y = rep(1:4, 3)[1:10]

cbind(x, y, i(x, y, 1))

#>              x y      x:y::2   x:y::3     x:y::4
#> 1   0.35859089 1  0.00000000 0.000000  0.0000000
#> 2  -0.02899503 2 -0.02899503 0.000000  0.0000000
#> 3   1.14704207 3  0.00000000 1.147042  0.0000000
#> 4   0.37358894 4  0.00000000 0.000000  0.3735889
#> 5   0.32333921 1  0.00000000 0.000000  0.0000000
#> 6  -0.82981932 2 -0.82981932 0.000000  0.0000000
#> 7   1.39446258 3  0.00000000 1.394463  0.0000000
#> 8  -0.19154358 4  0.00000000 0.000000 -0.1915436
#> 9   0.27227702 1  0.00000000 0.000000  0.0000000
#> 10 -1.08165901 2 -1.08165901 0.000000  0.0000000

#
# In fixest estimations
#

# NOTA: in fixest estimations, i(var, fe, ref) is equivalent to var::fe(ref)

data(base_did)
# We interact the variable 'period' with the variable 'treat'
est_did = feols(y ~ x1 + i(treat, period, 5) | id+period, base_did)

# You could have used the following formula instead:
# y ~ x1 + treat::period(5) | id+period