Prints the number of unique elements

This utility tool displays the number of unique elements in a data.frame as well as their number of NA values.

n_unik(x, fml)

# S3 method for vec_n_unik
print(x, ...)

Arguments

x
A data.frame or a vector.

fml
A one-sided formula containing the variables for which to count the number of unique values. You can use .N to get the number of observations. You can combine variables with the "^" operator. For example: fml = ~ .N + id + id^period will report the number of observations, the number of unique values of id and the number of unique id x period pairs. You can sub select variables with [], like in id[!is.na(period)]. You can use the special functions sw and sw0, like in ~id^sw0(period) + period[sw0(is.na(id))] which would lead to ~id + id^period + period + period[is.na(id)].

...
Not currently used.

Value

It returns a vector containing the number of unique values per element.

Examples

data = base_did
data$x1.L1 = round(lag(x1~id+period, 1, data))

# By default, just the formatted number of observations
n_unik(data)

#> #> # Observations: 1,080 

# Or the nber of unique elements of a vector
n_unik(data$id)

#> #> id: 108 

# number of unique id values and id x period pairs
n_unik(data, ~.N + id + id^period)

#> #> # Observations: 1,080 
#> #>             id:   108 
#> #>      id^period: 1,080 

# using sub selection + sw0
n_unik(data, ~.N + period[sw0(!is.na(x1.L1))])

#> #>        # Observations: 1,080 
#> #>                period:    10 
#> #> period[!is.na(x1.L1)]:     9