Finds observations to be removed from ML estimation with factors/clusters

For Poisson, Negative Binomial or Logit estimations with fixed-effects, when the dependent variable is only equal to 0 (or 1 for Logit) for one cluster value this leads to a perfect fit for that cluster value by setting its associated cluster coefficient to -Inf. Thus these observations need to be removed before estimation. This function gives the observations to be removed. Note that by default the function femlm or feglm drops them before performing the estimation.

obs2remove(fml, data, family = c("poisson", "negbin", "logit"))

Arguments

fml
A formula containing the dependent variable and the clusters. It can be of the type: y ~ cluster_1 + cluster_2 or y ~ x1 | cluster_1 + cluster_1 (in which case variables before the pipe are ignored).

data
A data.frame containing the variables in the formula.

family
Character scalar: either “poisson” (default), “negbin” or “logit”.

Value

It returns an integer vector of observations to be removed. If no observations are to be removed, an empty integer vector is returned. In both cases, it is of class fixest.obs2remove. The vector has an attribute cluster which is a list giving the IDs of the clusters that have been removed, for each cluster dimension.

Examples

base = iris
# v6: Petal.Length with only 0 values for 'setosa'
base$v6 = base$Petal.Length
base$v6[base$Species == "setosa"] = 0

(x = obs2remove(v6 ~ Species, base))

#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> attr(,"class")
#> [1] "fixest.obs2remove"
#> attr(,"family")
#> [1] "poisson"
#> attr(,"cluster")
#> attr(,"cluster")$Species
#> [1] "setosa"
#> 

attr(x, "cluster")

#> $Species
#> [1] "setosa"
#> 

# The two results are identical:
res_1 = femlm(v6 ~ Petal.Width | Species, base)

#> NOTE: 1 fixed-effect (50 observations) removed because of only zero outcomes.

# => note + obsRemoved is created

res_2 = femlm(v6 ~ Petal.Width | Species, base[-x, ])
# => no note because observations are removed before

esttable(res_1, res_2)

#>                               res_1           res_2
#> Petal.Width         0.2089 (0.1318) 0.2089 (0.1318)
#> Fixed-Effects:      --------------- ---------------
#> Species                         Yes             Yes
#> ___________________ _______________ _______________
#> Observations                    100             100
#> S.E. type            1-way: Species  1-way: Species
#> Squared Corr.                 0.699           0.699
#> Pseudo R2                   0.02726         0.02726
#> BIC                          376.54          376.54

all(res_1$obsRemoved == x)

#> [1] TRUE