Method to print the type of small sample correction

Pretty print for the small sample correction obtained from ssc

Usage

# S3 method for class 'ssc_type'
print(x, ...)

Arguments

x

An object of class ssc_type, obtained from the function ssc.

...

Not currently used.

Value

This function does not return anything.

See also

Examples

# ssc() is just a list
unclass(ssc())
#> $K.adj
#> [1] TRUE
#> 
#> $K.fixef
#> [1] "nonnested"
#> 
#> $G.adj
#> [1] TRUE
#> 
#> $G.df
#> [1] "min"
#> 
#> $t.df
#> [1] "min"
#> 
#> $K.exact
#> [1] FALSE
#> 

# print of ssc, much more compactly
ssc()
#> Small sample correction:
#> K.adj = TRUE, K.fixef = "nonnested", K.exact = FALSE
#> G.adj = TRUE, G.df = "min", t.df = "min"