Replicates fixest objects

Simple function that replicates fixest objects while (optionally) computing different standard-errors. Useful mostly in combination with etable or coefplot.

# S3 method for fixest
rep(x, times = 1, each = 1, vcov, ...)

# S3 method for fixest_list
rep(x, times = 1, each = 1, vcov, ...)

.l(...)

Arguments

x
Either a fixest object, either a list of fixest objects created with .l().

times
Integer vector giving the number of repetitions of the vector of elements. By default times = 1. It must be either of length 1, either of the same length as the argument x.

each
Integer scalar indicating the repetition of each element. Default is 1.

vcov
A list containing the types of standard-error to be computed, default is missing. If not missing, it must be of the same length as times, each, or the final vector. Note that if the arguments times and each are missing, then times becomes equal to the length of vcov. To see how to summon a VCOV, see the dedicated section in the vignette.

...
In .l(): fixest objects. In rep(): not currently used.

Value

Returns a list of the appropriate length. Each element of the list is a fixest object.

Details

To apply rep.fixest on a list of fixest objects, it is absolutely necessary to use .l() and not list().

Examples

# Let's show results with different standard-errors

est = feols(Ozone ~ Solar.R + Wind + Temp, data = airquality)
#> NOTE: 42 observations removed because of NA values (LHS: 37, RHS: 7).

my_vcov = list(~ Month, ~ Day, ~ Day + Month)

etable(rep(est, vcov = my_vcov))
#> Error: in etable(rep(est, vcov = my_vcov)):
#>  Some elements in '...' could not be evaluated: 
#>   in rep.fixest(est, vcov = my_vcov):
#>  'vcov' is not a valid argument of function rep.fixest (fyi, some of its
#> main arguments are 'times' and 'each').

coefplot(rep(est, vcov = my_vcov), drop = "Int")
#> Error: in rep.fixest(est, vcov = my_vcov):
#>  'vcov' is not a valid argument of function rep.fixest (fyi, some of its
#> main arguments are 'times' and 'each').

#
# To rep multiple objects, you need to use .l()
#

est_bis = feols(Ozone ~ Solar.R + Wind + Temp | Month, airquality)
#> NOTE: 42 observations removed because of NA values (LHS: 37, RHS: 7).

etable(rep(.l(est, est_bis), vcov = my_vcov))
#> Warning: le nombre d'objets à remplacer n'est pas multiple de la taille du remplacement
#> Warning: le nombre d'objets à remplacer n'est pas multiple de la taille du remplacement
#> Warning: Conversion automatique de LHS en liste
#> Warning: Conversion automatique de LHS en liste
#> Error in etable(rep(.l(est, est_bis), vcov = my_vcov)): The first element of '...' is not valid: it should be a fixest object or a list of fixest objects, it is neither.

# using each
etable(rep(.l(est, est_bis), each = 3, vcov = my_vcov))
#> Warning: le nombre d'objets à remplacer n'est pas multiple de la taille du remplacement
#> Warning: le nombre d'objets à remplacer n'est pas multiple de la taille du remplacement
#> Warning: Conversion automatique de LHS en liste
#> Warning: Conversion automatique de LHS en liste
#> Warning: Conversion automatique de LHS en liste
#> Warning: Conversion automatique de LHS en liste
#> Error in etable(rep(.l(est, est_bis), each = 3, vcov = my_vcov)): The first element of '...' is not valid: it should be a fixest object or a list of fixest objects, it is neither.