Sets the default type of standard errors to be used

This functions defines or extracts the default type of standard-errors to computed in fixest summary, and vcov.

setFixest_se(
  no_FE = "standard",
  one_FE = "cluster",
  two_FE = "cluster",
  all = NULL,
  reset = FALSE
)

getFixest_se()

Arguments

no_FE
Character scalar equal to either: "standard" (default), or "hetero". The type of standard-errors to use by default for estimations without fixed-effects.

one_FE
Character scalar equal to either: "standard", "hetero", or "cluster" (default). The type of standard-errors to use by default for estimations with one fixed-effect.

two_FE
Character scalar equal to either: "standard", "hetero", "cluster" (default), or "twoway". The type of standard-errors to use by default for estimations with two or more fixed-effects.

all
Character scalar equal to either: "standard", or "hetero". By default is is NULL. If provided, it sets all the SEs to that value.

reset
Logical, default is FALSE. Whether to reset to the default values.

Value

The function getFixest_se() returns a list with three elements containing the default for estimations i) without, ii) with one, or iii) with two or more fixed-effects.

Examples

# By default:
# - no fixed-effect (FE): standard
# - one or more FEs: cluster

data(base_did)
est_no_FE  = feols(y ~ x1, base_did)
est_one_FE = feols(y ~ x1 | id, base_did)
est_two_FE = feols(y ~ x1 | id + period, base_did)

etable(est_no_FE, est_one_FE, est_two_FE)

#>                          est_no_FE         est_one_FE         est_two_FE
#> Dependent Var.:                  y                  y                  y
#>                                                                         
#> (Intercept)      1.989*** (0.1492)                                      
#> x1              0.9831*** (0.0501) 0.9615*** (0.0481) 0.9779*** (0.0458)
#> Fixed-Effects:  ------------------ ------------------ ------------------
#> id                              No                Yes                Yes
#> period                          No                 No                Yes
#> _______________ __________________ __________________ __________________
#> S.E. type                 Standard             by: id             by: id
#> Observations                 1,080              1,080              1,080
#> R2                         0.26304            0.38715            0.48696
#> Within R2                       --            0.26507            0.30776

# Changing the default standard-errors
setFixest_se(no_FE = "hetero", one_FE = "standard", two_FE = "twoway")
etable(est_no_FE, est_one_FE, est_two_FE)

#>                          est_no_FE         est_one_FE         est_two_FE
#> Dependent Var.:                  y                  y                  y
#>                                                                         
#> (Intercept)      1.989*** (0.1490)                                      
#> x1              0.9831*** (0.0510) 0.9615*** (0.0514) 0.9779*** (0.0342)
#> Fixed-Effects:  ------------------ ------------------ ------------------
#> id                              No                Yes                Yes
#> period                          No                 No                Yes
#> _______________ __________________ __________________ __________________
#> S.E. type       Heteroskedas.-rob.           Standard    by: id & period
#> Observations                 1,080              1,080              1,080
#> R2                         0.26304            0.38715            0.48696
#> Within R2                       --            0.26507            0.30776

# Resetting the defaults
setFixest_se(reset = TRUE)