Cleans a character vector from multiple patterns

Recursively cleans a character vector from several patterns. Quickly handle the tedious task of data cleaning by taking advantage of the syntax. You can also apply all sorts of cleaning operations by summoning str_ops() operations.

str_clean(
  x,
  ...,
  replacement = "",
  pipe = " => ",
  sep = ",[ \n\t]+",
  ignore.case = FALSE,
  fixed = FALSE,
  word = FALSE,
  total = FALSE
)

Arguments

x

A character vector.

...

Character scalars representing patterns. A pattern is of the form "flags/pat1, pat2 => replacement". This means that patterns 'pat1' and 'pat2' will be replaced with the string 'replacement'. By default patterns are comma separated and the replacement comes after a ' => ' (see args sep and pipe to change this). By default the replacement is the empty string (so "pat1, pat2" removes the patterns).

Available flags are: 'word' (add word boundaries), 'ignore' (the case), 'fixed' (no regex), and 'total'. The flag total leads to a total replacement of the string if the pattern is found. Use flags with comma separation ("word, total/pat") or use only their initials ("wt/pat").

Starting with an '@' leads to operations in str_ops(). Ex: "@ascii, lower, ws" turns the string into ASCII, lowers the case and normalizes white spaces (see help of str_ops()).

replacement

Character scalar, default is the empty string. It represents the default value by which the patterns found in the character strings will be replaced. For example str_clean(x, "e", replacement = "a") turn all letters "e" in x into "a".

pipe

Character scalar, default is " => ". If thevalue of pipe is found in a pattern, then the string is split w.r.t. the pipe and anything after the pipe becomes the replacement.

For example in str_clean(x, "e => a") the default pipe is found in "e => a", so the pattern "e" will be replaced with "a". In other terms, this is equivalent to str_clean(x, "e", replacement = "a"). Example changing the pipe: you can obtain the previous result with str_clean(x, "e|>a", pipe = "|>").

sep

Character scalar, default is ",[ \t\n]+" (which means a comma followed with spaces and/or new lines). By default the patterns to be replaced are comma separated, that is the pattern is split w.r.t. the argument sep and a replacement is done for each sub-pattern.

Use NULL or the empty string to disable pattern separation.

For example: let's look at str_clean(x, "w/one, two => three"). First the flag "word" is extracted from the pattern (see arg. ...) as well as the replacement (see arg. pipe), leading to "one, two" the pattern to be replaced. Then the pattern is split w.r.t. sep, leading to two patterns "one" and "two". Hence the words (thanks to the flag "w") "one" and "two" from the string x will be replaced with "three".

ignore.case

Logical scalar, default is FALSE. If TRUE, then case insensitive search is triggered.

fixed

Logical scalar, default is FALSE. Whether to trigger a fixed search instead of a regular expression search (default).

word

Logical scalar, default is FALSE. If TRUE then a) word boundaries are added to the pattern, and b) patterns can be chained by separating them with a comma, they are combined with an OR logical operation. Example: if word = TRUE, then pattern = "The, mountain" will select strings containing either the word 'The' or the word 'mountain'.

total

Logical scalar, default is FALSE. If TRUE, then when a pattern is found in a string, the full string is replaced (instead of just the pattern). Note, importantly, that when total = TRUE you can use logical operators in the patterns.

Example: str_clean(x, "wi/ & two, three & !four => ", total = TRUE)

Value

The main usage returns a character vector of the same length as the vector in input. Note, however, that since you can apply arbitrary str_ops() operations, the length and type of the final vector may depend on those (if they are used).

See also

Author

Laurent R. Berge

Examples

x = c("hello world  ", "it's 5 am....")

# we clean the o's and the points (we use 'fixed' to trigger fixed-search)
str_clean(x, "o", "f/.")
#> [1] "hell wrld  " "it's 5 am"  
# equivalently
str_clean(x, "fixed/o, .")
#> [1] "hell wrld  " "it's 5 am"  
# equivalently
str_clean(x, "o, .", fixed = TRUE)
#> [1] "hell wrld  " "it's 5 am"  
# equivalently
str_clean(x, "o", ".", fixed = TRUE)
#> [1] "hell wrld  " "it's 5 am"  

#
# chaining operations: example using cars
#

cars = row.names(mtcars)
new = str_clean(cars, 
           # replace strings containing "Maz" with Mazda
           "total/Maz => Mazda", 
           # replace the word 'Merc' with Mercedes
           "wi/merc => Mercedes",
           # replace strings containing "Merc" and a digit followed with an 'S'
           "t/Merc & \\dS => Mercedes S!",
           # put to lower case, remove isolated characters and normalize white spaces
           "@lower, ws.isolated")

cbind(cars, new)
#>       cars                  new                  
#>  [1,] "Mazda RX4"           "mazda"              
#>  [2,] "Mazda RX4 Wag"       "mazda"              
#>  [3,] "Datsun 710"          "datsun 710"         
#>  [4,] "Hornet 4 Drive"      "hornet drive"       
#>  [5,] "Hornet Sportabout"   "hornet sportabout"  
#>  [6,] "Valiant"             "valiant"            
#>  [7,] "Duster 360"          "duster 360"         
#>  [8,] "Merc 240D"           "mercedes 240d"      
#>  [9,] "Merc 230"            "mercedes 230"       
#> [10,] "Merc 280"            "mercedes 280"       
#> [11,] "Merc 280C"           "mercedes 280c"      
#> [12,] "Merc 450SE"          "mercedes s!"        
#> [13,] "Merc 450SL"          "mercedes s!"        
#> [14,] "Merc 450SLC"         "mercedes s!"        
#> [15,] "Cadillac Fleetwood"  "cadillac fleetwood" 
#> [16,] "Lincoln Continental" "lincoln continental"
#> [17,] "Chrysler Imperial"   "chrysler imperial"  
#> [18,] "Fiat 128"            "fiat 128"           
#> [19,] "Honda Civic"         "honda civic"        
#> [20,] "Toyota Corolla"      "toyota corolla"     
#> [21,] "Toyota Corona"       "toyota corona"      
#> [22,] "Dodge Challenger"    "dodge challenger"   
#> [23,] "AMC Javelin"         "amc javelin"        
#> [24,] "Camaro Z28"          "camaro z28"         
#> [25,] "Pontiac Firebird"    "pontiac firebird"   
#> [26,] "Fiat X1-9"           "fiat x1-9"          
#> [27,] "Porsche 914-2"       "porsche 914-2"      
#> [28,] "Lotus Europa"        "lotus europa"       
#> [29,] "Ford Pantera L"      "ford pantera"       
#> [30,] "Ferrari Dino"        "ferrari dino"       
#> [31,] "Maserati Bora"       "maserati bora"      
#> [32,] "Volvo 142E"          "volvo 142e"