Efficient creation of string vectors with optional interpolation

Create string vectors in multiple ways: 1) add successive string elements (like in c()), or 2) write a character string that will be broken with respect to commas ("hi, there" becomes c("hi", "there")), or 3) interpolate variables in character strings ("x{1:2}" becomes c("x1", "x2")) with full access to smagick() operations, or any combination of the three.

str_vec(
  ...,
  .delim = c("{", "}"),
  .envir = parent.frame(),
  .split = TRUE,
  .protect.vars = TRUE
)

Arguments

...

Character vectors that will be vectorized. If commas are present in the character vector, it will be split with respect to commas and following blanks. The vectors can contain any interpolation in the form "{var}" and any smagick() operation can be applied. To change the delimiters for interpolation, see .delim. Named arguments are used in priority for variable substitution, otherwise the value of the variables to be interpolated are fetched in the calling environment (see argument .envir).

Note, importantly, that interpolation and comma splitting are performed on "natural" vectors only. That is: str_vec("x{1:5}") will lead to a vector of length 5 ("x1" to "x5"), while z = "x{1:5}" followed by str_vec(z) leads to a vector of length 1: "x{1:5}". To change this behavior and obtain equivalent results, use .protect.vars = FALSE.

.delim

Character vector of length 1 or 2. Default is c("{", "}"). Defines the opening and the closing delimiters for interpolation.

If of length 1, it must be of the form: 1) the opening delimiter, 2) a single space, 3) the closing delimiter. Ex: ".[ ]" is equivalent to c(".[", "]"). The default value is equivalent to "{ }".

[ ]: R:%20 [", "]: R:%22,%20%22

.envir

An environment used to evaluate the variables in "{}". By default the variables are evaluated using the environment from where the function is called or using the named arguments passed to the function.

.split

Logical, default is TRUE. Whether to split the vector with respect to commas. Ex: by default str_vec("hi, there") leads to the vector c("hi", "there").

.protect.vars

Logical scalar, default is TRUE. If TRUE, then only arguments equal to a "natural" character scalar are comma-split and interpolated, other arguments are not touched. Ex: str_vec("x{1:5}") will lead to a vector of length 5 ("x1" to "x5"), while z = "x{1:5}" followed by str_vec(z) leads to a vector of length 1: "x{1:5}". If FALSE, comma-splitting and interpolation is performed on all variables.

Details

The default of the argument .protect.vars is FALSE so as to avoid unwanted comma-splitting and interpolations. The main use case of this function is the creation of small string vectors, which can be written directly at function call.

See also

Author

Laurent Berge

Examples

# illustrating comma-splitting and interpolation
str_vec("x1, y2, z{4:5}")
#> [1] "x1" "y2" "z4" "z5"

# variable protection
x = "x{1:5}"
str_vec(x, "y{1:2}")
#> [1] "x{1:5}" "y1"     "y2"    

# without protection => interpolation takes place
str_vec(x, "y{1:2}", .protect.vars = FALSE)
#> [1] "x1" "x2" "x3" "x4" "x5" "y1" "y2"

# removing comma splitting
str_vec("Hi, said Charles.", "Hi, said {girl}.", girl = "Julia", .split = FALSE)
#> [1] "Hi, said Charles." "Hi, said Julia."  

# changing the delimiters for interpolation
pkg = "\\usepackage[usenames,dvipsnames]{xcolor}"
str_vec("\\usepackage{.[S!graphicx, fourier, standalone]}", 
        pkg, .delim = ".[ ]")
#> [1] "\\usepackage{graphicx}"                   
#> [2] "\\usepackage{fourier}"                    
#> [3] "\\usepackage{standalone}"                 
#> [4] "\\usepackage[usenames,dvipsnames]{xcolor}"