Method to subselect from a fixest_panel

Subselection from a fixest_panel which has been created with the function panel. Also allows to create lag/lead variables with functions l()/f() if the fixest_panel is also a data.table.

# S3 method for fixest_panel
[(x, i, j, ...)

Arguments

x
A fixest_panel object, created with the function panel.

i
Row subselection. Allows data.table style selection (provided the data is also a data.table).

j
Variable selection. Allows data.table style selection/variable creation (provided the data is also a data.table).

...
Other arguments to be passed to [.data.frame or data.table (or whatever the class of the initial data).

Value

It returns a fixest_panel data base, with the attributes allowing to create lags/leads properly bookkeeped.

Details

If the original data was also a data.table, some calls to [.fixest_panel may dissolve the fixest_panel object and return a regular data.table. This is the case for subselections with additional arguments. If so, a note is displayed on the console.

See also

Examples

data(base_did)

# Creating a fixest_panel object
pdat = panel(base_did, ~id+period)

# Subselections of fixest_panel objects bookkeeps the leads/lags engine
pdat_small = pdat[!pdat$period %in% c(2, 4), ]
a = feols(y~l(x1, 0:1), pdat_small)

#> Error in feols(y ~ l(x1, 0:1), pdat_small): The right hand side contains NA values (e.g. observations 1, 2, 3, 9, 10, 319 others and NA). Please provide data without NA values (or use na_inf.rm). FYI the variables (Intercept), x1 and NA are concerned.

# we obtain the same results, had we created the lags "on the fly"
base_small = base_did[!base_did$period %in% c(2, 4), ]
b = feols(y~l(x1, 0:1), base_small, panel.id = ~id+period)

#> Error in feols(y ~ l(x1, 0:1), base_small, panel.id = ~id + period): The right hand side contains NA values (e.g. observations 1, 2, 3, 9, 10, 319 others and NA). Please provide data without NA values (or use na_inf.rm). FYI the variables (Intercept), x1 and NA are concerned.

etable(a, b)

#> Error in results2formattedList(..., se = se, dof = dof, fitstat = fitstat,     cluster = cluster, digits = digits, sdBelow = sdBelow, signifCode = signifCode,     title = title, subtitles = subtitles, yesNoFixef = yesNoFixef,     keepFactors = keepFactors, isTex = tex, useSummary = useSummary,     dots_call = dots_call, powerBelow = powerBelow, dict = dict,     interaction.combine = interaction.combine, convergence = convergence,     show_family = family, drop = drop, file = file, order = order,     label = label, fixef_sizes = fixef_sizes, show_depvar = depvar): objet 'a' introuvable

# \donttest{
# Using data.table to create new lead/lag variables
pdat_dt = panel(as.data.table(base_did), ~id+period)

# Variable creation
pdat_dt[, x_l1 := l(x1)]

#> Error in `[.data.table`(x = pdat_dt, j = `:=`(x_l1, l(x1))): objet 'pdat_dt' introuvable

pdat_dt[, c("x_l1", "x_f1_2") := .(l(x1), f(x1)**2)]

#> Error in `[.data.table`(x = pdat_dt, j = `:=`(c("x_l1", "x_f1_2"), .(l(x1),     f(x1)^2))): objet 'pdat_dt' introuvable

# Estimation on a subset of the data
#  (the lead/lags work appropriately)
feols(y~l(x1, 0:1), pdat_dt[!period %in% c(2, 4)])

#> Error in feols(y ~ l(x1, 0:1), pdat_dt[!period %in% c(2, 4)]): The right hand side contains NA values (e.g. observations 1, 2, 3, 9, 10, 319 others and NA). Please provide data without NA values (or use na_inf.rm). FYI the variables (Intercept), x1 and NA are concerned.

# }