Summary of a fixest object. Computes different types of standard errors.

This function is similar to print.fixest. It provides the table of coefficients along with other information on the fit of the estimation. It can compute different types of standard errors. The new variance covariance matrix is an object returned.

# S3 method for fixest
summary(object, se, cluster, dof = getFixest_dof(),
  forceCovariance = FALSE, keepBounded = FALSE, ...)

summ(object, se, cluster, dof = getFixest_dof(),
  forceCovariance = FALSE, keepBounded = FALSE, ...)

Arguments

object
A fixest object. Obtained using the functions femlm, feols or feglm.

se
Character scalar. Which kind of standard error should be computed: “standard”, “White”, “cluster”, “twoway”, “threeway” or “fourway”? By default if there are clusters in the estimation: se = "cluster", otherwise se = "standard". Note that this argument can be implicitly deduced from the argument cluster.

cluster
Tells how to cluster the standard-errors (if clustering is requested). Can be either a list of vectors, a character vector of variable names, a formula or an integer vector. Assume we want to perform 2-way clustering over var1 and var2 contained in the data.frame base used for the estimation. All the following cluster arguments are valid and do the same thing: cluster = base[, c("var1, "var2")]}, \code{cluster = c("var1, "var2"), cluster = ~var1+var2. If the two variables were used as clusters in the estimation, you could further use cluster = 1:2 or leave it blank with se = "twoway" (assuming var1 [resp. var2] was the 1st [res. 2nd] cluster).

dof
An object of class dof.type obtained with the function dof. Represent how the degree of freedom correction should be done. Defaults to dof(fixef="nested", exact=FALSE, cluster = TRUE). See the help of the function dof for details.

forceCovariance
(Advanced users.) Logical, default is FALSE. In the peculiar case where the obtained Hessian is not invertible (usually because of collinearity of some variables), use this option to force the covariance matrix, by using a generalized inverse of the Hessian. This can be useful to spot where possible problems come from.

keepBounded
(Advanced users -- feNmlm with non-linear part and bounded coefficients only.) Logical, default is FALSE. If TRUE, then the bounded coefficients (if any) are treated as unrestricted coefficients and their S.E. is computed (otherwise it is not).

...
Not currently used.

Value

It returns a fixest object with:

cov.scaled

The new variance-covariance matrix (computed according to the argument se).

se

The new standard-errors (computed according to the argument se).

coeftable

The table of coefficients with the new standard errors.

See also

Examples

# Load trade data
data(trade)

# We estimate the effect of distance on trade (with 3 cluster effects)
est_pois = femlm(Euros ~ log(dist_km)|Origin+Destination+Product, trade)

# Comparing different types of standard errors
sum_white    = summary(est_pois, se = "white")
sum_oneway   = summary(est_pois, se = "cluster")
sum_twoway   = summary(est_pois, se = "twoway")
sum_threeway = summary(est_pois, se = "threeway")

esttable(sum_white, sum_oneway, sum_twoway, sum_threeway)

#>                              sum_white          sum_oneway          sum_twoway
#> log(dist_km)        -1.5277*** (0.022) -1.5277*** (0.1156) -1.5277*** (0.1323)
#> Fixed-Effects:      ------------------ ------------------- -------------------
#> Origin                             Yes                 Yes                 Yes
#> Destination                        Yes                 Yes                 Yes
#> Product                            Yes                 Yes                 Yes
#> ___________________ __________________ ___________________ ___________________
#> Observations                    38,325              38,325              38,325
#> S.E. type                        White       1-way: Origin 2-way: Orig. & De..
#> Squared Corr.                    0.604               0.604               0.604
#> Pseudo R2                      0.76039             0.76039             0.76039
#> BIC                          1.427e+12           1.427e+12           1.427e+12
#>                            sum_threeway
#> log(dist_km)        -1.5277*** (0.1369)
#> Fixed-Effects:      -------------------
#> Origin                              Yes
#> Destination                         Yes
#> Product                             Yes
#> ___________________ ___________________
#> Observations                     38,325
#> S.E. type           3-way: Orig. & De..
#> Squared Corr.                     0.604
#> Pseudo R2                       0.76039
#> BIC                           1.427e+12

# Alternative ways to cluster the SE:
# \donttest{
# two-way clustering: Destination and Product
# (Note that arg. se = "twoway" is implicitly deduced from the argument cluster)
summary(est_pois, cluster = c("Destination", "Product"))

#> ML estimation, family = Poisson, Dep. Var.: Euros
#> Observations: 38,325 
#> Fixed-effects: Origin: 15,  Destination: 15,  Product: 20
#> Standard-errors: Two-way (Destination & Product) 
#>              Estimate Std. Error z value  Pr(>|z|)    
#> log(dist_km)  -1.5277   0.120086 -12.722 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> Log-likelihood: -7.133e+11   Adj. Pseudo-R2: 0.76039 
#>            BIC:  1.427e+12     Squared Cor.: 0.60377 

summary(est_pois, cluster = trade[, c("Destination", "Product")])

#> ML estimation, family = Poisson, Dep. Var.: Euros
#> Observations: 38,325 
#> Fixed-effects: Origin: 15,  Destination: 15,  Product: 20
#> Standard-errors: Two-way 
#>              Estimate Std. Error z value  Pr(>|z|)    
#> log(dist_km)  -1.5277   0.120086 -12.722 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> Log-likelihood: -7.133e+11   Adj. Pseudo-R2: 0.76039 
#>            BIC:  1.427e+12     Squared Cor.: 0.60377 

summary(est_pois, cluster = list(trade$Destination, trade$Product))

#> ML estimation, family = Poisson, Dep. Var.: Euros
#> Observations: 38,325 
#> Fixed-effects: Origin: 15,  Destination: 15,  Product: 20
#> Standard-errors: Two-way 
#>              Estimate Std. Error z value  Pr(>|z|)    
#> log(dist_km)  -1.5277   0.120086 -12.722 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> Log-likelihood: -7.133e+11   Adj. Pseudo-R2: 0.76039 
#>            BIC:  1.427e+12     Squared Cor.: 0.60377 

summary(est_pois, cluster = ~Destination+Product)

#> ML estimation, family = Poisson, Dep. Var.: Euros
#> Observations: 38,325 
#> Fixed-effects: Origin: 15,  Destination: 15,  Product: 20
#> Standard-errors: Two-way (Destination & Product) 
#>              Estimate Std. Error z value  Pr(>|z|)    
#> log(dist_km)  -1.5277   0.120086 -12.722 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> Log-likelihood: -7.133e+11   Adj. Pseudo-R2: 0.76039 
#>            BIC:  1.427e+12     Squared Cor.: 0.60377 

# Since Destination and Product are used as fixed-effects, you can also use:
summary(est_pois, cluster = 2:3)

#> ML estimation, family = Poisson, Dep. Var.: Euros
#> Observations: 38,325 
#> Fixed-effects: Origin: 15,  Destination: 15,  Product: 20
#> Standard-errors: Two-way (Destination & Product) 
#>              Estimate Std. Error z value  Pr(>|z|)    
#> log(dist_km)  -1.5277   0.120086 -12.722 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> Log-likelihood: -7.133e+11   Adj. Pseudo-R2: 0.76039 
#>            BIC:  1.427e+12     Squared Cor.: 0.60377 

# }