Summary method for fixed-effects coefficients

This function summarizes the main characteristics of the fixed-effects coefficients. It shows the number of fixed-effects that have been set as references and the first elements of the fixed-effects.

Usage

# S3 method for fixest.fixef
summary(object, n = 5, ...)

Arguments

object

An object returned by the function fixef.fixest.

n

Positive integer, defaults to 5. The n first fixed-effects for each fixed-effect dimension are reported.

...

Not currently used.

Value

It prints the number of fixed-effect coefficients per fixed-effect dimension, as well as the number of fixed-effects used as references for each dimension, and the mean and variance of the fixed-effect coefficients. Finally, it reports the first 5 (arg. n) elements of each fixed-effect.

See also

Author

Laurent Berge

Examples

data(trade)

# We estimate the effect of distance on trade
# => we account for 3 fixed-effects effects
est_pois = femlm(Euros ~ log(dist_km)|Origin+Destination+Product, trade)

# obtaining the fixed-effects coefficients
fe_trade = fixef(est_pois)

# printing some summary information on the fixed-effects coefficients:
summary(fe_trade)
#> Fixed_effects coefficients
#>                         Origin Destination Product
#> Number of fixed-effects     15          15      20
#> Number of references         0           1       1
#> Mean                      23.5        3.09  0.0127
#> Standard-deviation        1.28        1.11    1.36
#> 
#> COEFFICIENTS:
#>   Origin:    AT    BE    DE   DK    ES                 
#>           22.67 23.72 24.87 23.6 25.13 ... 10 remaining
#> -----
#>   Destination:    AT    BE    DE    DK    ES                 
#>                2.436 2.696 4.323 2.451 4.043 ... 10 remaining
#> -----
#>   Product: 1     2     3     4      5                 
#>            0 1.414 0.656 1.449 -1.521 ... 15 remaining