Extracts the variance/covariance of a femlm fit

This function extracts the variance-covariance of estimated parameters from a model estimated with femlm, feols or feglm.

# S3 method for fixest
vcov(object, se, cluster, dof = getFixest_dof(),
  forceCovariance = FALSE, keepBounded = FALSE, ...)

Arguments

object
A fixest object. Obtained using the functions femlm, feols or feglm.

se
Character scalar. Which kind of standard error should be computed: “standard”, “White”, “cluster”, “twoway”, “threeway” or “fourway”? By default if there are clusters in the estimation: se = "cluster", otherwise se = "standard". Note that this argument can be implicitly deduced from the argument cluster.

cluster
Tells how to cluster the standard-errors (if clustering is requested). Can be either a list of vectors, a character vector of variable names, a formula or an integer vector. Assume we want to perform 2-way clustering over var1 and var2 contained in the data.frame base used for the estimation. All the following cluster arguments are valid and do the same thing: cluster = base[, c("var1, "var2")]}, \code{cluster = c("var1, "var2"), cluster = ~var1+var2. If the two variables were used as clusters in the estimation, you could further use cluster = 1:2 or leave it blank with se = "twoway" (assuming var1 [resp. var2] was the 1st [res. 2nd] cluster).

dof
An object of class dof.type obtained with the function dof. Represent how the degree of freedom correction should be done. Defaults to dof(fixef="nested", exact=FALSE, cluster = TRUE). See the help of the function dof for details.

forceCovariance
(Advanced users.) Logical, default is FALSE. In the peculiar case where the obtained Hessian is not invertible (usually because of collinearity of some variables), use this option to force the covariance matrix, by using a generalized inverse of the Hessian. This can be useful to spot where possible problems come from.

keepBounded
(Advanced users -- feNmlm with non-linear part and bounded coefficients only.) Logical, default is FALSE. If TRUE, then the bounded coefficients (if any) are treated as unrestricted coefficients and their S.E. is computed (otherwise it is not).

...
Other arguments to be passed to summary.fixest.

The computation of the VCOV matrix is first done in summary.fixest.

Value

It returns a \(N\times N\) square matrix where \(N\) is the number of variables of the fitted model. This matrix has an attribute “type” specifying how this variance/covariance matrix has been computed (i.e. was it created using White correction, or was it clustered along a specific factor, etc).

See also

Examples

# Load trade data
data(trade)

# We estimate the effect of distance on trade (with 3 fixed-effects)
est_pois = femlm(Euros ~ log(dist_km) + log(Year) | Origin + Destination +
                 Product, trade)

# By default, in the presence of FEs
# the VCOV is clustered along the first FE
vcov(est_pois)

#>            [,1]       [,2]
#> [1,] 0.01337806  0.2270412
#> [2,] 0.22704123 48.7565447
#> attr(,"type")
#> [1] "Clustered (Origin)"
#> attr(,"dof.type")
#> [1] "dof(fixef = \"nested\", exact = FALSE, cluster = TRUE)"
#> attr(,"dof.K")
#> [1] 36

# "white" VCOV
vcov(est_pois, se = "white")

#>              [,1]         [,2]
#> [1,] 0.0004774988  0.002270112
#> [2,] 0.0022701117 39.998250222
#> attr(,"type")
#> [1] "White"
#> attr(,"dof.type")
#> [1] "dof(fixef = \"nested\", exact = FALSE, cluster = TRUE)"
#> attr(,"dof.K")
#> [1] 50

# "clustered" VCOV (with respect to the Product factor)
vcov(est_pois, se = "cluster", cluster = trade$Product)

#>            [,1]        [,2]
#> [1,] 0.00965938  0.09050064
#> [2,] 0.09050064 45.72458483
#> attr(,"type")
#> [1] "Clustered"
#> attr(,"dof.type")
#> [1] "dof(fixef = \"nested\", exact = FALSE, cluster = TRUE)"
#> attr(,"dof.K")
#> [1] 31

# another way to make the same request:
# note that previously arg. se was optional since deduced from arg. cluster
vcov(est_pois, cluster = "Product")

#>            [,1]        [,2]
#> [1,] 0.00965938  0.09050064
#> [2,] 0.09050064 45.72458483
#> attr(,"type")
#> [1] "Clustered (Product)"
#> attr(,"dof.type")
#> [1] "dof(fixef = \"nested\", exact = FALSE, cluster = TRUE)"
#> attr(,"dof.K")
#> [1] 31

# yet another way:
vcov(est_pois, cluster = ~Product)

#>            [,1]        [,2]
#> [1,] 0.00965938  0.09050064
#> [2,] 0.09050064 45.72458483
#> attr(,"type")
#> [1] "Clustered (Product)"
#> attr(,"dof.type")
#> [1] "dof(fixef = \"nested\", exact = FALSE, cluster = TRUE)"
#> attr(,"dof.K")
#> [1] 31

# Another estimation without cluster:
est_pois_simple = femlm(Euros ~ log(dist_km) + log(Year), trade)

# We can still get the clustered VCOV,
# but we need to give the argument cluster:
vcov(est_pois_simple, cluster = ~Product)

#>              [,1]         [,2]          [,3]
#> [1,] 2590.3952791 -0.362100706 -339.93823080
#> [2,]   -0.3621007  0.002112443    0.04584858
#> [3,] -339.9382308  0.045848576   44.61192316
#> attr(,"type")
#> [1] "Clustered (Product)"
#> attr(,"dof.type")
#> [1] "dof(fixef = \"nested\", exact = FALSE, cluster = TRUE)"
#> attr(,"dof.K")
#> [1] 3