Estimations table (export the results of multiples estimations to a DF or to Latex)

Aggregates the results of multiple estimations and displays them in the form of either a Latex table or a data.frame.

etable(..., se = c("standard", "white", "cluster", "twoway", "threeway",
  "fourway"), dof = getFixest_dof(), cluster, digits = 4, tex, fitstat,
  title, sdBelow = TRUE, drop, order, dict = getFixest_dict(), file,
  replace = FALSE, convergence, signifCode, label, subtitles,
  fixef_sizes = FALSE, yesNoFixef = c("Yes", "No"),
  keepFactors = TRUE, family, powerBelow = -5,
  interaction.combine = " $\\times $ ", depvar)

esttex(..., se = c("standard", "white", "cluster", "twoway", "threeway",
  "fourway"), dof = getFixest_dof(), cluster, digits = 4, fitstat,
  title, sdBelow = TRUE, drop, order, dict = getFixest_dict(), file,
  replace = FALSE, convergence, signifCode = c(`***` = 0.01, `**` =
  0.05, `*` = 0.1), label, subtitles, fixef_sizes = FALSE,
  yesNoFixef = c("Yes", "No"), keepFactors = TRUE, family,
  powerBelow = -5, interaction.combine = " $\\times $ ")

esttable(..., se = c("standard", "white", "cluster", "twoway",
  "threeway", "fourway"), dof = getFixest_dof(), cluster, depvar, drop,
  order, digits = 4, fitstat, convergence, signifCode = c(`***` =
  0.001, `**` = 0.01, `*` = 0.05, . = 0.1), subtitles,
  keepFactors = FALSE, family)

Arguments

...
Used to capture different fixest estimation objects (obtained with femlm, feols or feglm). Note that any other type of element is discarded. Note that you can give a list of fixest objects.

se
Character scalar. Which kind of standard error should be computed: “standard”, “White”, “cluster”, “twoway”, “threeway” or “fourway”? By default if there are clusters in the estimation: se = "cluster", otherwise se = "standard". Note that this argument can be implicitly deduced from the argument cluster.

dof
An object of class dof.type obtained with the function dof. Represent how the degree of freedom correction should be done. Defaults to dof(fixef="nested", exact=FALSE, cluster = TRUE). See the help of the function dof for details.

cluster
Tells how to cluster the standard-errors (if clustering is requested). Can be either a list of vectors, a character vector of variable names, a formula or an integer vector. Assume we want to perform 2-way clustering over var1 and var2 contained in the data.frame base used for the estimation. All the following cluster arguments are valid and do the same thing: cluster = base[, c("var1, "var2")]}, \code{cluster = c("var1, "var2"), cluster = ~var1+var2. If the two variables were used as clusters in the estimation, you could further use cluster = 1:2 or leave it blank with se = "twoway" (assuming var1 [resp. var2] was the 1st [res. 2nd] cluster).

digits
Integer, default is 4. The number of digits to be displayed.

tex
Logical: whether the results should be a data.frame or a Latex table. By default, this argument is TRUE if the argument file (used for exportation) is not missing; it is equal to FALSE otherwise.

fitstat
A character vector or a one sided formula. A vector listing which fit statistics to display. The valid types are 'll', 'aic', 'bic' and r2 types like 'r2', 'pr2', 'war2', etc (see all valid types in r2). The default value depends on the models to display. Example of use: fitstat=c('sq.cor', 'ar2', 'war2'), or fitstat=~sq.cor+ar2+war2 using a formula.

title
(Tex only.) Character scalar. The title of the Latex table.

sdBelow
(Tex only.) Logical, default is TRUE. Should the standard-errors be displayed below the coefficients?

drop
Character vector. This element is used if some variables are not to be displayed. This should be a vector of regular expressions (see regex help for more info). Each variable satisfying any of the regular expressions will be discarded. This argument is applied post aliasing (see argument dict). Example: you have the variable x1 to x55 and want to display only x1 to x9, then you could use drop = "x[[:digit:]]{2}". If the first character is an exclamation mark, the effect is reversed (e.g. drop = "!Intercept" means: every variable that does not contain “Intercept” is dropped). See details.

order
Character vector. This element is used if the user wants the variables to be ordered in a certain way. This should be a vector of regular expressions (see regex help for more info). The variables satisfying the first regular expression will be placed first, then the order follows the sequence of regular expressions. This argument is applied post aliasing (see argument dict). Example: you have the following variables: month1 to month6, then x1 to x5, then year1 to year6. If you want to display first the x's, then the years, then the months you could use: order = c("x", "year"). If the first character is an exclamation mark, the effect is reversed (e.g. order = "!Intercept" means: every variable that does not contain “Intercept” goes first). See details.

dict
(Tex only.) A named character vector. It changes the original variable names to the ones contained in the dict. E.g. to change the variables named a and b3 to (resp.) “$log(a)$” and to “$bonus^3$”, use dict=c(a="$log(a)$",b3="$bonus^3$"). By default it is equal to getFixest_dict(), a default dictionary which can be set with setFixest_dict.

file
A character scalar. If provided, the Latex (or data frame) table will be saved in a file whose path is file. If you provide this argument, then a Latex table will be exported, to export a regular data.frame, use argument tex = FALSE.

replace
Logical, default is FALSE. Only used if option file is used. Should the exported table be written in a new file that replaces any existing file?

convergence
Logical, default is missing. Should the convergence state of the algorithm be displayed? By default, convergence information is displayed if at least one model did not converge.

signifCode
Named numeric vector, used to provide the significance codes with respect to the p-value of the coefficients. Default is c("***"=0.01, "**"=0.05, "*"=0.10) for a Latex table and c("***"=0.001, "**"=0.01, "*"=0.05, "."=0.10) for a data.frame (to conform with R's default).

label
(Tex only.) Character scalar. The label of the Latex table.

subtitles
Character vector of the same length as the number of models to be displayed. If provided, subtitles are added underneath the dependent variable name.

fixef_sizes
(Tex only.) Logical, default is FALSE. If TRUE and fixed-effects were used in the models, then the number of "individuals" per fixed-effect dimension is also displayed.

yesNoFixef
(Tex only.) A character vector of length 1 or 2. Default is c("Yes", "No"). This is the message displayed when a given cluster is (or is not) included in a regression. If yesNoFixef is of length 1, then the second element is the empty string.

keepFactors
Logical, default is TRUE. If FALSE, then factor variables are displayed as fixed-effects and no coefficient is shown.

family
Logical, default is missing. Whether to display the families of the models. By default this line is displayed when at least two models are from different families.

powerBelow
(Tex only.) Integer, default is -5. A coefficient whose value is below 10**(powerBelow+1) is written with a power in Latex. For example 0.0000456 would be written 4.56$\times 10^{-5}$ by default. Setting powerBelow = -6 would lead to 0.00004 in Latex.

interaction.combine
(Tex only.) Character scalar, defaults to " $\times$ ". When the estimation contains interactions, then the variables names (after aliasing) are combined with this argument. For example: if dict = c(x1="Wind", x2="Rain") and you have the following interaction x1:x2, then it will be renamed (by default) Wind $\times$ Rain -- using interaction.combine = "*" would lead to Wind*Rain.

depvar
(Data frame only.) Logical, default is missing. Whether a first line containing the dependent variables should be shown. By default, the dependent variables are shown only if they differ across models or if the argumen file is not missing.

Value

If tex = TRUE, the lines composing the Latex table are returned invisibly while the table is directly prompted on the console.

If tex = FALSE, the data.frame is directly returned. If the argument file is not missing, the data.frame is returned invisibly.

Details

The function esttex is equivalent to the function etable with argument tex = TRUE.

The function esttable is equivalent to the function etable with argument tex = FALSE.

Functions

esttex: Exports the results of multiple fixest estimations in a Latex table.

esttable: Facility to display the results of multiple fixest estimations.

Arguments drop and order

The arguments drop and order use regular expressions. If you are aware of regular expressions, I urge you to learn it, since it is an extremely powerful way to manipulate character strings (and it exists across most programming languages).

For example drop = "Wind" would drop any variable whose name contains "Wind". Note that variables such as "Temp:Wind" or "StrongWind" do contain "Wind", so would be dropped. To drop only the variable named "Wind", you need to use drop = "^Wind$" (with "^" meaning beginning, resp. "$" meaning end, of the string => this is the language of regular expressions).

Although you can combine several regular expressions in a single character string using pipes, drop also accepts a vector of regular expressions.

You can use the special character "!" (exclamation mark) to reverse the effect of the regular expression (this feature is specific to this fonction). For example drop = "!Wind" would drop any variable that does not contain "Wind".

The argument order takes in a vector of regular expressions, the order will follow the elments of this vector. The vector gives a list of priorities, on the left the elements with highest priority. For example, order = c("Wind", "!Inter", "!Temp") would give highest priorities to the variables containing "Wind" (which would then appear first), second highest priority is the variables not containing "Inter", last, with lowest priority, the variables not containing "Temp". If you had the following variables: (Intercept), Temp:Wind, Wind, Temp you would end up with the following order: Wind, Temp:Wind, Temp, (Intercept).

See also

Examples

aq = airquality
aq$Month = factor(aq$Month)

est1 = feols(Ozone ~ Month / Wind + Temp, data = aq)

#> NOTE: 37 observations removed because of NA values (Breakup: LHS: 37, RHS: 0).

est2 = feols(Ozone ~ Wind + Temp | Month, data = aq)

#> NOTE: 37 observations removed because of NA values (Breakup: LHS: 37, RHS: 0, Clusters: 0).

# Displaying the two results in a single table
etable(est1, est2)

#>                                      est1               est2
#> (Intercept)         -100.236*** (26.2326)                   
#> Month6                -54.9912* (26.2302)                   
#> Month7                 35.8878. (18.2884)                   
#> Month8                  44.5181* (17.971)                   
#> Month9                 -11.7802 (18.1378)                   
#> Temp                   2.0424*** (0.3065) 2.1049*** (0.1939)
#> Month5:Wind              -1.0857 (1.1219)                   
#> Month6:Wind               2.0456 (1.7761)                   
#> Month7:Wind           -5.6156*** (1.3104)                   
#> Month8:Wind            -6.515*** (1.2148)                   
#> Month9:Wind              -1.3487 (1.1417)                   
#> Wind                                       -2.7817* (1.2629)
#> Fixed-Effects:      --------------------- ------------------
#> Month                                  No                Yes
#> ___________________ _____________________ __________________
#> Observations                          116                116
#> S.E. type                        Standard       1-way: Month
#> R2                                0.68106            0.60155
#> Within R2                              --              0.479

# drop: dropping the non interaction terms (see regexp help)
etable(est1, est2, drop = "^[[:alnum:]]+$")

#>                                      est1    est2
#> (Intercept)         -100.236*** (26.2326)        
#> Month5:Wind              -1.0857 (1.1219)        
#> Month6:Wind               2.0456 (1.7761)        
#> Month7:Wind           -5.6156*** (1.3104)        
#> Month8:Wind            -6.515*** (1.2148)        
#> Month9:Wind              -1.3487 (1.1417)        
#> Fixed-Effects:      ---------------------      --
#> Month                                  No     Yes
#> ___________________ _____________________      __
#> Observations                          116     116
#> S.E. type                        Standard      ..
#> R2                                0.68106 0.60155
#> Within R2                              --   0.479

# drop interactions
etable(est1, est2, drop = ":")

#>                                      est1               est2
#> (Intercept)         -100.236*** (26.2326)                   
#> Month6                -54.9912* (26.2302)                   
#> Month7                 35.8878. (18.2884)                   
#> Month8                  44.5181* (17.971)                   
#> Month9                 -11.7802 (18.1378)                   
#> Temp                   2.0424*** (0.3065) 2.1049*** (0.1939)
#> Wind                                       -2.7817* (1.2629)
#> Fixed-Effects:      --------------------- ------------------
#> Month                                  No                Yes
#> ___________________ _____________________ __________________
#> Observations                          116                116
#> S.E. type                        Standard       1-way: Month
#> R2                                0.68106            0.60155
#> Within R2                              --              0.479

# keep only interactions ("!" reverses the effect)
etable(est1, est2, drop = "!:")

#>                                    est1    est2
#> Month5:Wind            -1.0857 (1.1219)        
#> Month6:Wind             2.0456 (1.7761)        
#> Month7:Wind         -5.6156*** (1.3104)        
#> Month8:Wind          -6.515*** (1.2148)        
#> Month9:Wind            -1.3487 (1.1417)        
#> Fixed-Effects:      -------------------      --
#> Month                                No     Yes
#> ___________________ ___________________      __
#> Observations                        116     116
#> S.E. type                      Standard      ..
#> R2                              0.68106 0.60155
#> Within R2                            --   0.479

# order: Wind variable first, intercept last
etable(est1, est2, order = c("Wind", "Month"))

#>                                      est1               est2
#> Month5:Wind              -1.0857 (1.1219)                   
#> Month6:Wind               2.0456 (1.7761)                   
#> Month7:Wind           -5.6156*** (1.3104)                   
#> Month8:Wind            -6.515*** (1.2148)                   
#> Month9:Wind              -1.3487 (1.1417)                   
#> Wind                                       -2.7817* (1.2629)
#> Month6                -54.9912* (26.2302)                   
#> Month7                 35.8878. (18.2884)                   
#> Month8                  44.5181* (17.971)                   
#> Month9                 -11.7802 (18.1378)                   
#> (Intercept)         -100.236*** (26.2326)                   
#> Temp                   2.0424*** (0.3065) 2.1049*** (0.1939)
#> Fixed-Effects:      --------------------- ------------------
#> Month                                  No                Yes
#> ___________________ _____________________ __________________
#> Observations                          116                116
#> S.E. type                        Standard       1-way: Month
#> R2                                0.68106            0.60155
#> Within R2                              --              0.479

etable(est1, est2, order = c("^Wind", "Wind", "Month"))

#>                                      est1               est2
#> Wind                                       -2.7817* (1.2629)
#> Month5:Wind              -1.0857 (1.1219)                   
#> Month6:Wind               2.0456 (1.7761)                   
#> Month7:Wind           -5.6156*** (1.3104)                   
#> Month8:Wind            -6.515*** (1.2148)                   
#> Month9:Wind              -1.3487 (1.1417)                   
#> Month6                -54.9912* (26.2302)                   
#> Month7                 35.8878. (18.2884)                   
#> Month8                  44.5181* (17.971)                   
#> Month9                 -11.7802 (18.1378)                   
#> (Intercept)         -100.236*** (26.2326)                   
#> Temp                   2.0424*** (0.3065) 2.1049*** (0.1939)
#> Fixed-Effects:      --------------------- ------------------
#> Month                                  No                Yes
#> ___________________ _____________________ __________________
#> Observations                          116                116
#> S.E. type                        Standard       1-way: Month
#> R2                                0.68106            0.60155
#> Within R2                              --              0.479

# Interactions, then Intercept, last ("!" reverses the effect)
etable(est1, est2, order = c("!Int", "!:"))

#>                                      est1               est2
#> Month6                -54.9912* (26.2302)                   
#> Month7                 35.8878. (18.2884)                   
#> Month8                  44.5181* (17.971)                   
#> Month9                 -11.7802 (18.1378)                   
#> Temp                   2.0424*** (0.3065) 2.1049*** (0.1939)
#> Wind                                       -2.7817* (1.2629)
#> Month5:Wind              -1.0857 (1.1219)                   
#> Month6:Wind               2.0456 (1.7761)                   
#> Month7:Wind           -5.6156*** (1.3104)                   
#> Month8:Wind            -6.515*** (1.2148)                   
#> Month9:Wind              -1.3487 (1.1417)                   
#> (Intercept)         -100.236*** (26.2326)                   
#> Fixed-Effects:      --------------------- ------------------
#> Month                                  No                Yes
#> ___________________ _____________________ __________________
#> Observations                          116                116
#> S.E. type                        Standard       1-way: Month
#> R2                                0.68106            0.60155
#> Within R2                              --              0.479

# signifCode
etable(est1, est2, signifCode = c(" A"=0.01, " B"=0.05, " C"=0.1,
                                  " D"=0.15, " F"=1))

#>                                     est1               est2
#> (Intercept)         -100.236 A (26.2326)                   
#> Month6              -54.9912 B (26.2302)                   
#> Month7               35.8878 C (18.2884)                   
#> Month8                44.5181 B (17.971)                   
#> Month9              -11.7802 F (18.1378)                   
#> Temp                   2.0424 A (0.3065)  2.1049 A (0.1939)
#> Month5:Wind           -1.0857 F (1.1219)                   
#> Month6:Wind            2.0456 F (1.7761)                   
#> Month7:Wind           -5.6156 A (1.3104)                   
#> Month8:Wind            -6.515 A (1.2148)                   
#> Month9:Wind           -1.3487 F (1.1417)                   
#> Wind                                     -2.7817 B (1.2629)
#> Fixed-Effects:      -------------------- ------------------
#> Month                                 No                Yes
#> ___________________ ____________________ __________________
#> Observations                         116                116
#> S.E. type                       Standard       1-way: Month
#> R2                               0.68106            0.60155
#> Within R2                             --              0.479

# fitstat
etable(est1, est2, fitstat = ~r2+ar2+apr2+war2)

#>                                      est1               est2
#> (Intercept)         -100.236*** (26.2326)                   
#> Month6                -54.9912* (26.2302)                   
#> Month7                 35.8878. (18.2884)                   
#> Month8                  44.5181* (17.971)                   
#> Month9                 -11.7802 (18.1378)                   
#> Temp                   2.0424*** (0.3065) 2.1049*** (0.1939)
#> Month5:Wind              -1.0857 (1.1219)                   
#> Month6:Wind               2.0456 (1.7761)                   
#> Month7:Wind           -5.6156*** (1.3104)                   
#> Month8:Wind            -6.515*** (1.2148)                   
#> Month9:Wind              -1.3487 (1.1417)                   
#> Wind                                       -2.7817* (1.2629)
#> Fixed-Effects:      --------------------- ------------------
#> Month                                  No                Yes
#> ___________________ _____________________ __________________
#> Observations                          116                116
#> S.E. type                        Standard       1-way: Month
#> R2                                0.68106            0.60155
#> Adjusted R2                       0.65069            0.57962
#> Adj. Pseudo R2                     0.0988            0.08316
#> Within Adj. R2                         --            0.46944

# Adding a dictionnary (Tex only)
dict = c(Month5="May", Month6="Jun", Month7="Jul", Month8="Aug", Month9="Sep")
etable(est1, est2, dict = dict, tex = TRUE)

#> \begin{table}[htbp]\centering
#> \caption{no title}
#> \begin{tabular}{lcc}
#>  & & \tabularnewline
#> \hline
#> \hline
#> Dependent Variable:&\multicolumn{2}{c}{Ozone}\\
#> Model:&(1)&(2)\\
#> \hline
#> \emph{Variables}\tabularnewline
#> (Intercept)&-100.24$^{***}$&  \\
#>   &(26.233)&  \\
#> Jun&-54.991$^{**}$&  \\
#>   &(26.230)&  \\
#> Jul&35.888$^{*}$&  \\
#>   &(18.288)&  \\
#> Aug&44.518$^{**}$&  \\
#>   &(17.971)&  \\
#> Sep&-11.780&  \\
#>   &(18.138)&  \\
#> Temp&2.0424$^{***}$&2.1049$^{***}$\\
#>   &(0.3065)&(0.1939)\\
#> May $\times $ Wind&-1.0857&  \\
#>   &(1.1219)&  \\
#> Jun $\times $ Wind&2.0456&  \\
#>   &(1.7761)&  \\
#> Jul $\times $ Wind&-5.6156$^{***}$&  \\
#>   &(1.3104)&  \\
#> Aug $\times $ Wind&-6.5150$^{***}$&  \\
#>   &(1.2148)&  \\
#> Sep $\times $ Wind&-1.3487&  \\
#>   &(1.1417)&  \\
#> Wind&  &-2.7817$^{**}$\\
#>   &  &(1.2629)\\
#> \hline
#> \emph{Fixed-Effects}&  & \\
#> Month&No&Yes\\
#> \hline
#> \emph{Fit statistics}&  & \\
#> Observations& 116&116\\
#> Standard-Error type& Standard&One-way: Month\\
#> R$^2$ & 0.68106&0.60155\\
#> Within R$^2$ & --&0.479\\
#> \hline
#> \hline
#> \multicolumn{3}{l}{\emph{Signif Codes: ***: 0.01, **: 0.05, *: 0.1}}\\
#> \end{tabular}
#> \end{table}