Extract the formula of a fixest fit

This function extracts the formula from a fixest estimation (obtained with femlm, feols or feglm). If the estimation was done with fixed-effects, they are added in the formula after a pipe (“|”). If the estimation was done with a non linear in parameters part, then this will be added in the formula in between I().

Usage

# S3 method for class 'fixest'
formula(x, type = c("full", "linear", "iv", "NL"), ...)

Arguments

x

An object of class fixest. Typically the result of a femlm, feols or feglm estimation.

type

A character scalar. Default is type = "full" which gives back a formula containing the linear part of the model along with the fixed-effects (if any) and the IV part (if any). If type = "linear" then only the linear formula is returned. If type = "NL" then only the non linear in parameters part is returned.

...

Not currently used.

Value

It returns a formula.

See also

Author

Laurent Berge

Examples

# simple estimation on iris data, using "Species" fixed-effects
res = femlm(Sepal.Length ~ Sepal.Width + Petal.Length +
            Petal.Width | Species, iris)

# formula with the fixed-effect variable
formula(res)
#> Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width | Species

# linear part without the fixed-effects
formula(res, "linear")
#> Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width
#> <environment: 0x0000017ad952c190>