Hat values for fixest objects

Computes the hat values for feols or feglm estimations. Only works when there are no fixed-effects.

Usage

# S3 method for fixest
hatvalues(model, ...)

Arguments

model

A fixest object. For instance from feols or feglm.

...

Not currently used.

Value

Returns a vector of the same length as the number of observations used in the estimation.

Details

Hat values are not available for fenegbin, femlm and feNmlm estimations.

When there are fixed-effects, the hat values of the reduced form are different from the hat values of the full model. And we cannot get costlessly the hat values of the full model from the reduced form. It would require to reestimate the model with the fixed-effects as regular variables.

Examples

est = feols(Petal.Length ~ Petal.Width + Sepal.Width, iris)
head(hatvalues(est))
#> [1] 0.02042587 0.02110119 0.01838332 0.01933430 0.02273855 0.03265226