Create, or interact variables with, factors

Treat a variable as a factor, or interacts a variable with another treated as a factor. Values to be dropped/kept from the factor can be easily set.

i(var, f, f2, ref, drop, keep, drop2, keep2)

interact

Arguments

var
A vector to be interacted with f. If the other argument f is missing, then this vector will be treated as the argument f.

f
A vector (of any type) that will be treated as a factor. Must be of the same length as var if var is not missing.

f2
A vector (of any type) that will be treated as a factor. Must be of the same length as f.

ref
A single value that belongs to the interacted variable (f). Can be missing, can also be a logical: if TRUE, then the first value of f will be removed..

drop
A vector of regular expressions or integers (if f is integer). If provided, all values from f that match drop will be removed.

keep
A vector of regular expressions or integers (if f is integer). If provided, only the values from f that match keep will be kept.

drop2
A vector of regular expressions or integers (if f2 is integer). If provided, all values from f2 that match drop2 will be removed.

keep2
A vector of regular expressions or integers (if f2 is integer). If provided, only the values from f2 that match keep2 will be kept.

Format

An object of class function of length 1.

Value

It returns a matrix with number of rows the length of var. The number of columns is equal to the number of cases contained in f minus the reference(s).

Shorthand in fixest estimations

In fixest estimations, instead of using i(var, f, ref), you can instead use the following writing var::f(ref). Note that this way of doing interactions is not endorsed any more and will likely be deprecated in the future.

See also

Author

Laurent Berge

Examples

#
# Simple illustration
#

x = 1:10
y = rep(1:4, 3)[1:10]

# interaction
cbind(x, y, i(x, y, 1))

#> Error in as.character(x): cannot coerce type 'closure' to vector of type 'character'

# without interaction
cbind(x, y, i(y, ref = 1))

#> Error in as.character(x): cannot coerce type 'closure' to vector of type 'character'

# You can interact factors too
z = rep(c("a", "b", "c"), c(5, 3, 2))
data.frame(z, y, i(z, y))

#> Error in as.character(x): cannot coerce type 'closure' to vector of type 'character'

#
# In fixest estimations
#

data(base_did)
# We interact the variable 'period' with the variable 'treat'
est_did = feols(y ~ x1 + i(treat, period, 5) | id + period, base_did)

#> Error in feols(y ~ x1 + i(treat, period, 5) | id + period, base_did): Evaluation of the right-hand-side of the formula raises an error: cannot
#> coerce type 'closure' to vector of type 'character'

# => special treatment in coefplot
coefplot(est_did)

#> Error in coefplot_prms(object = object, ..., sd = sd, ci_low = ci_low,     ci_high = ci_high, x = x, x.shift = x.shift, dict = dict,     keep = keep, drop = drop, order = order, ci_level = ci_level,     ref = ref, only.inter = only.inter, sep = sep, as.multiple = as.multiple): object 'est_did' not found

# Using i() for factors
est_bis = feols(y ~ x1 + i(period, keep = 3:6) + i(treat, period, 5) | id, base_did)

#> Error in feols(y ~ x1 + i(period, keep = 3:6) + i(treat, period, 5) |     id, base_did): Evaluation of the right-hand-side of the formula raises an error: cannot
#> coerce type 'closure' to vector of type 'character'

coefplot(est_bis, only.inter = FALSE)

#> Error in coefplot_prms(object = object, ..., sd = sd, ci_low = ci_low,     ci_high = ci_high, x = x, x.shift = x.shift, dict = dict,     keep = keep, drop = drop, order = order, ci_level = ci_level,     ref = ref, only.inter = only.inter, sep = sep, as.multiple = as.multiple): object 'est_bis' not found

# => special treatment in etable
etable(est_bis, dict = c("6" = "six"))

#> Error: in etable(est_bis, dict = c(`6` = "six")):
#>  Some elements in '...' could not be evaluated: object 'est_bis' not
#> found

#
# Interact two factors => f2
#

# To interact two factor, use the argument f2
data(airquality)
aq = airquality
aq$week = aq$Day %/% 7 + 1

# Interacting Month and week:
res_2F = feols(Ozone ~ Solar.R + i(Month, f2 = week), aq)

#> Error in feols(Ozone ~ Solar.R + i(Month, f2 = week), aq): Evaluation of the right-hand-side of the formula raises an error: cannot
#> coerce type 'closure' to vector of type 'character'

# Same but dropping the 5th Month and 1st week
res_2F_bis = feols(Ozone ~ Solar.R + i(Month, f2 = week, drop = 5, drop2 = 1), aq)

#> Error in feols(Ozone ~ Solar.R + i(Month, f2 = week, drop = 5, drop2 = 1),     aq): Evaluation of the right-hand-side of the formula raises an error: cannot
#> coerce type 'closure' to vector of type 'character'

etable(res_2F, res_2F_bis)

#> Error: in etable(res_2F, res_2F_bis):
#>  Some elements in '...' could not be evaluated: object 'res_2F' not found