Lags a variable in a fixest estimation

Produce lags or leads in the formulas of fixest estimations or when creating variables in a data.table. The data must be set as a panel beforehand (either with the function panel or with the argument panel.id in the estimation).

f(x, lead = 1, fill = NA)

l(x, lag = 1, fill = NA)

Arguments

x
The variable.

lead
A vector of integers giving the number of leads. Negative values lead to lags. This argument can be a vector when using it in fixest estimations. When creating variables in a data.table, it **must** be of length one.

fill
A scalar, default is NA. How to fill the missing values due to the lag/lead? Note that in a fixest estimation, 'fill' must be numeric (not required when creating new variables).

lag
A vector of integers giving the number of lags. Negative values lead to leads. This argument can be a vector when using it in fixest estimations. When creating variables in a data.table, it **must** be of length one.

Value

These functions can only be used i) in a formula of a fixest estimation, or ii) when creating variables within a fixest_panel object (obtained with function panel) which is alaos a data.table.

Functions

f: Forwards a variable (inverse of lagging) in a fixest estimation

See also

Examples

data(base_did)

# Setting a data set as a panel...
pdat = panel(base_did, ~id+period)

# ...then using the functions l and f
est1 = feols(y~l(x1, 0:1), pdat)

#> NOTE: 108 observations removed because of NA values (Breakup: LHS: 0, RHS: 108).

est2 = feols(f(y)~l(x1, -1:1), pdat)

#> NOTE: 216 observations removed because of NA values (Breakup: LHS: 108, RHS: 216).

est3 = feols(l(y)~l(x1, 0:3), pdat)

#> NOTE: 324 observations removed because of NA values (Breakup: LHS: 108, RHS: 324).

etable(est1, est2, est3, order = c("f", "^x"), drop="Int")

#>                                   est1              est2               est3
#> Dependent Var.:                      y            f(y,1)             l(y,1)
#>                                                                            
#> f(x1,1)                                0.994*** (0.0578)                   
#> x1                  0.9948*** (0.0531)   0.0081 (0.0584)   -0.0534 (0.0599)
#> l(x1,1)                  0.041 (0.054)   0.0157 (0.0585) 0.9871*** (0.0613)
#> l(x1,2)                                                      0.022 (0.0607)
#> l(x1,3)                                                     0.0102 (0.0598)
#> ___________________ __________________ _________________ __________________
#> Observations                       972               864                756
#> S.E. type                     Standard          Standard           Standard
#> R2                             0.26558           0.25697            0.25875
#> Adjusted R2                    0.26406           0.25438             0.2548

# or using the argument panel.id
feols(f(y)~l(x1, -1:1), base_did, panel.id = ~id+period)

#> NOTE: 216 observations removed because of NA values (Breakup: LHS: 108, RHS: 216).

#> OLS estimation, Dep. Var.: f(y, 1)
#> Observations: 864 
#> Standard-errors: Standard 
#>             Estimate Std. Error   z value  Pr(>|z|)    
#> (Intercept) 2.464300   0.169611 14.529000 < 2.2e-16 ***
#> f(x1, 1)    0.994018   0.057828 17.189000 < 2.2e-16 ***
#> x1          0.008072   0.058366  0.138297  0.890038    
#> l(x1, 1)    0.015693   0.058506  0.268223  0.788592    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> Log-likelihood: -2,612.04   Adj. R2: 0.25438 

# \donttest{
# l() and f() can also be used within a data.table:
pdat_dt = panel(as.data.table(base_did), ~id+period)

#> Error in as.data.table(base_did): impossible de trouver la fonction "as.data.table"

# Now since pdat_dt is also a data.table
#   you can create lags/leads directly
pdat_dt[, x1_l1 := l(x1)]

#> Error in eval(expr, envir, enclos): objet 'pdat_dt' introuvable

pdat_dt[, c("x1_l1_fill0", "y_f2") := .(l(x1, fill = 0), f(y, 2))]

#> Error in eval(expr, envir, enclos): objet 'pdat_dt' introuvable

# }