Finds observations to be removed from ML estimation with fixed-effects

For Poisson, Negative Binomial or Logit estimations with fixed-effects, when the dependent variable is only equal to 0 (or 1 for Logit) for one fixed-effect value this leads to a perfect fit for that fixed-effect value by setting its associated fixed-effect coefficient to -Inf. Thus these observations need to be removed before estimation. This function gives the observations to be removed. Note that by default the function femlm or feglm drops them before performing the estimation.

obs2remove(fml, data, family = c("poisson", "negbin", "logit"))

Arguments

fml
A formula containing the dependent variable and the fixed-effects. It can be of the type: y ~ fixef_1 + fixef_2 or y ~ x1 | fixef_1 + fixef_1 (in which case variables before the pipe are ignored).

data
A data.frame containing the variables in the formula.

family
Character scalar: either “poisson” (default), “negbin” or “logit”.

Value

It returns an integer vector of observations to be removed. If no observations are to be removed, an empty integer vector is returned. In both cases, it is of class fixest.obs2remove. The vector has an attribute fixef which is a list giving the IDs of the fixed-effects that have been removed, for each fixed-effect dimension.

See also

Examples

base = iris
# v6: Petal.Length with only 0 values for 'setosa'
base$v6 = base$Petal.Length
base$v6[base$Species == "setosa"] = 0

(x = obs2remove(v6 ~ Species, base))

#> Warning: This function is deprecated and will disappear in 1 year from 12/11/2020.

#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> attr(,"class")
#> [1] "fixest.obs2remove"
#> attr(,"family")
#> [1] "poisson"
#> attr(,"fixef")
#> attr(,"fixef")$v6
#> [1] 0
#> 
#> attr(,"fixef")$Species
#> [1] "setosa"
#> 

attr(x, "fixef")

#> $v6
#> [1] 0
#> 
#> $Species
#> [1] "setosa"
#> 

# The two results are identical:
res_1 = femlm(v6 ~ Petal.Width | Species, base)

#> NOTE: 1 fixed-effect (50 observations) removed because of only 0 outcomes.

# => note + obsRemoved is created

res_2 = femlm(v6 ~ Petal.Width | Species, base[-x, ])
# => no note because observations are removed before

esttable(res_1, res_2)

#> Warning: The argument 'subtitles' is deprecated. Please use 'headers' instead.

#>                           res_1           res_2
#> Dependent Var.:              v6              v6
#>                                                
#> Petal.Width     0.2089 (0.1324) 0.2089 (0.1324)
#> Fixed-Effects:  --------------- ---------------
#> Species                     Yes             Yes
#> _______________ _______________ _______________
#> S.E.: Clustered     by: Species     by: Species
#> Observations                100             100
#> Squared Cor.            0.69884         0.69884
#> Pseudo R2               0.02726         0.02726
#> BIC                      362.73          362.73

all(res_1$obs_selection$obsRemoved == x)

#> [1] FALSE