A print facility for fixest objects.

This function is very similar to usual summary functions as it provides the table of coefficients along with other information on the fit of the estimation. The type of output is customizable by the user (using function setFixest_print.type).

# S3 method for fixest
print(x, n, type = getFixest_print.type(), ...)

Arguments

x
A fixest object. Obtained using the methods femlm, feols or feglm.

n
Integer, number of coefficients to display. By default, only the first 8 coefficients are displayed if x does not come from summary.fixest.

type
Either "table" (default) to display the coefficients table or "coef" to display only the coefficients. By default the value is getFixest_print.type() which can be permanently set with setFixest_print.type).

...
Other arguments to be passed to vcov.fixest.

See also

Examples

# Load trade data
data(trade)

# We estimate the effect of distance on trade
#   => we account for 3 fixed-effects (FEs)
est_pois = femlm(Euros ~ log(dist_km)|Origin+Destination+Product, trade)

# displaying the results
#  (by default SEs are clustered if FEs are used)
print(est_pois)

#> ML estimation, family = Poisson, Dep. Var.: Euros
#> Observations: 38,325 
#> Fixed-effects: Origin: 15,  Destination: 15,  Product: 20
#> Standard-errors: Clustered (Origin) 
#>              Estimate Std. Error z value  Pr(>|z|)    
#> log(dist_km)  -1.5277   0.115627 -13.213 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> Log-likelihood: -7.133e+11   Adj. Pseudo-R2: 0.76039 
#>            BIC:  1.427e+12     Squared Cor.: 0.60377 

# By default the coefficient table is displayed.
#  If the user wished to display only the coefficents, use option type:
print(est_pois, type = "coef")

#> log(dist_km) 
#>    -1.527747 

# \donttest{
# To permanently display coef. only, use setFixest_print.type:
setFixest_print.type("coef")
est_pois

#> log(dist_km) 
#>    -1.527747 

# back to default:
setFixest_print.type("table")
# }