Sets/gets the dictionary used in esttex

Sets/gets the default dictionary used in the function esttex. The dictionaries are used to relabel variables (usually towards a fancier, more explicit formatting) when exporting them into a Latex table. By setting the dictionary with setFixest_dict, you can avoid providing the argument dict in function esttex.

setFixest_dict(dict)

getFixest_dict()

Arguments

dict
A named character vector. E.g. to change my variable named "a" and "b" to (resp.) "$log(a)$" and "$bonus^3$", then use dict = c(a="$log(a)$", b3="$bonus^3$").

Examples

# \donttest{
data(trade)
est = feols(log(Euros) ~ log(dist_km)|Origin+Destination+Product, trade)
# we export the result & rename some variables
esttex(est, dict = c("log(Euros)"="Euros (ln)", Origin="Country of Origin"))

#> \begin{table}[htbp]\centering
#> \caption{no title}
#> \begin{tabular}{lc}
#>  & \tabularnewline
#> \hline
#> \hline
#> Dependent Variable:&Euros (ln)\\
#> Model:&(1)\\
#> \hline
#> \emph{Variables}\tabularnewline
#> log(dist\_km)&-2.1689$^{***}$\\
#>   &(0.1542)\\
#> \hline
#> \emph{Fixed-Effects}&  \\
#> Country of Origin&Yes\\
#> Destination&Yes\\
#> Product&Yes\\
#> \hline
#> \emph{Fit statistics}&  \\
#> Observations& 38,325\\
#> R$^2$ & 0.70402\\
#> Within R$^2$ & 0.21827\\
#> \hline
#> \hline
#> \multicolumn{2}{l}{\emph{One-way (Country of Origin) standard-errors in parentheses.}}\\
#> \multicolumn{2}{l}{\emph{Signif Codes: ***: 0.01, **: 0.05, *: 0.1}}\\
#> \end{tabular}
#> \end{table}

# If you export many tables, it can be more convenient to use setFixest_dict:
setFixest_dict(c("log(Euros)"="Euros (ln)", Origin="Country of Origin"))
esttex(est) # variables are properly relabeled

#> \begin{table}[htbp]\centering
#> \caption{no title}
#> \begin{tabular}{lc}
#>  & \tabularnewline
#> \hline
#> \hline
#> Dependent Variable:&Euros (ln)\\
#> Model:&(1)\\
#> \hline
#> \emph{Variables}\tabularnewline
#> log(dist\_km)&-2.1689$^{***}$\\
#>   &(0.1542)\\
#> \hline
#> \emph{Fixed-Effects}&  \\
#> Country of Origin&Yes\\
#> Destination&Yes\\
#> Product&Yes\\
#> \hline
#> \emph{Fit statistics}&  \\
#> Observations& 38,325\\
#> R$^2$ & 0.70402\\
#> Within R$^2$ & 0.21827\\
#> \hline
#> \hline
#> \multicolumn{2}{l}{\emph{One-way (Country of Origin) standard-errors in parentheses.}}\\
#> \multicolumn{2}{l}{\emph{Signif Codes: ***: 0.01, **: 0.05, *: 0.1}}\\
#> \end{tabular}
#> \end{table}

# }