Sets the default type of standard errors to be used

This functions defines or extracts the default type of standard-errors to computed in fixest summary, and vcov.

Usage

setFixest_vcov(
  no_FE = "iid",
  one_FE = "cluster",
  two_FE = "cluster",
  panel = "cluster",
  all = NULL,
  reset = FALSE
)

getFixest_vcov()

Arguments

no_FE

Character scalar equal to either: "iid" (default), or "hetero". The type of standard-errors to use by default for estimations without fixed-effects.

one_FE

Character scalar equal to either: "iid", "hetero", or "cluster" (default). The type of standard-errors to use by default for estimations with one fixed-effect.

two_FE

Character scalar equal to either: "iid", "hetero", "cluster" (default), or "twoway". The type of standard-errors to use by default for estimations with two or more fixed-effects.

panel

Character scalar equal to either: "iid", "hetero", "cluster" (default), or "driscoll_kraaay". The type of standard-errors to use by default for estimations with the argument panel.id set up. Note that panel has precedence over the presence of fixed-effects.

all

Character scalar equal to either: "iid", or "hetero". By default is is NULL. If provided, it sets all the SEs to that value.

reset

Logical, default is FALSE. Whether to reset to the default values.

Value

The function getFixest_vcov() returns a list with three elements containing the default for estimations i) without, ii) with one, or iii) with two or more fixed-effects.

Examples

# By default:
# - no fixed-effect (FE): standard
# - one or more FEs: cluster
# - panel: cluster on panel id

data(base_did)
est_no_FE  = feols(y ~ x1, base_did)
est_one_FE = feols(y ~ x1 | id, base_did)
est_two_FE = feols(y ~ x1 | id + period, base_did)
est_panel = feols(y ~ x1 | id + period, base_did, panel.id = ~id + period)

etable(est_no_FE, est_one_FE, est_two_FE)
#>                          est_no_FE         est_one_FE         est_two_FE
#> Dependent Var.:                  y                  y                  y
#>                                                                         
#> Constant         1.989*** (0.1492)                                      
#> x1              0.9831*** (0.0501) 0.9615*** (0.0481) 0.9779*** (0.0458)
#> Fixed-Effects:  ------------------ ------------------ ------------------
#> id                              No                Yes                Yes
#> period                          No                 No                Yes
#> _______________ __________________ __________________ __________________
#> S.E. type                      IID             by: id             by: id
#> Observations                 1,080              1,080              1,080
#> R2                         0.26304            0.38715            0.48696
#> Within R2                       --            0.26507            0.30776
#> ---
#> Signif. codes: 0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

# Changing the default standard-errors
setFixest_vcov(no_FE = "hetero", one_FE = "iid",
               two_FE = "twoway", panel = "drisc")
etable(est_no_FE, est_one_FE, est_two_FE, est_panel)
#>                          est_no_FE         est_one_FE         est_two_FE
#> Dependent Var.:                  y                  y                  y
#>                                                                         
#> Constant         1.989*** (0.1490)                                      
#> x1              0.9831*** (0.0510) 0.9615*** (0.0514) 0.9779*** (0.0342)
#> Fixed-Effects:  ------------------ ------------------ ------------------
#> id                              No                Yes                Yes
#> period                          No                 No                Yes
#> _______________ __________________ __________________ __________________
#> S.E. type       Heteroskedas.-rob.                IID    by: id & period
#> Observations                 1,080              1,080              1,080
#> R2                         0.26304            0.38715            0.48696
#> Within R2                       --            0.26507            0.30776
#> 
#>                          est_panel
#> Dependent Var.:                  y
#>                                   
#> Constant                          
#> x1              0.9779*** (0.0241)
#> Fixed-Effects:  ------------------
#> id                             Yes
#> period                         Yes
#> _______________ __________________
#> S.E. type       Drisco.-Kra. (L=1)
#> Observations                 1,080
#> R2                         0.48696
#> Within R2                  0.30776
#> ---
#> Signif. codes: 0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

# Resetting the defaults
setFixest_vcov(reset = TRUE)