Summary method for cluster coefficients

This function summarizes the main characteristics of the cluster coefficients. It shows the number of fixed-effects that have been set as references and the first elements of the fixed-effects.

# S3 method for fixest.fixef
summary(object, n = 5, ...)

Arguments

object
An object returned by the function fixef.fixest.

n
Positive integer, defaults to 5. The n first fixed-effects for each cluster are reported.

...
Not currently used.

Value

It prints the number of fixed-effect coefficients per cluster, as well as the number of fixed-effects used as references for each cluster, and the mean and variance of the cluster coefficients. Finally it reports the first 5 elements of each cluster.

See also

Examples

data(trade)

# We estimate the effect of distance on trade
# => we account for 3 fixed-effects effects
est_pois = femlm(Euros ~ log(dist_km)|Origin+Destination+Product, trade)

# obtaining the fixed-effects coefficients
fe_trade = fixef(est_pois)

# printing some summary information on the cluster coefficients:
summary(fe_trade)

#> Fixed_effects coefficients
#>                         Origin Destination Product
#> Number of fixed-effects     15          15      20
#> Number of references         0           1       1
#> Mean                      23.5        3.09  0.0127
#> Variance                  1.63        1.23    1.86
#> 
#> COEFFICIENTS:
#>   Origin:    SE    PT    LU    IT    IE                 
#>           23.41 22.61 20.39 24.49 23.36 ... 10 remaining
#> -----
#>   Destination:   SE    PT LU    IT    IE                 
#>                3.57 2.552  0 4.218 2.589 ... 10 remaining
#> -----
#>   Product: 1     2     3     4      5                 
#>            0 1.414 0.656 1.449 -1.521 ... 15 remaining