Summary for fixest_multi objects

Summary information for fixest_multi objects. In particular, this is used to specify the type of standard-errors to be computed.

# S3 method for fixest_multi
summary(
  object,
  type = "short",
  vcov = NULL,
  se = NULL,
  cluster = NULL,
  ssc = NULL,
  .vcov = NULL,
  stage = 2,
  lean = FALSE,
  n = 1000,
  ...
)

Arguments

object
A fixest_multi object, obtained from a fixest estimation leading to multiple results.

type
A character either equal to "short", "long", "compact", or "se_compact". If short, only the table of coefficients is displayed for each estimation. If long, then the full results are displayed for each estimation. If compact, a data.frame is returned with one line per model and the formatted coefficients + standard-errors in the columns. If se_compact, a data.frame is returned with one line per model, one numeric column for each coefficient and one numeric column for each standard-error.

vcov
Versatile argument to specify the VCOV. In general, it is either a character scalar equal to a VCOV type, either a formula of the form: vcov_type ~ variables. The VCOV types implemented are: "iid", "hetero" (or "HC1"), "cluster", "twoway", "NW" (or "newey_west"), "DK" (or "driscoll_kraay"), and "conley". It also accepts object from vcov_cluster, vcov_NW, NW, vcov_DK, DK, vcov_conley and conley. It also accepts covariance matrices computed externally. Finally it accepts functions to compute the covariances. See the `vcov` documentation in the vignette.

se
Character scalar. Which kind of standard error should be computed: “standard”, “hetero”, “cluster”, “twoway”, “threeway” or “fourway”? By default if there are clusters in the estimation: se = "cluster", otherwise se = "iid". Note that this argument is deprecated, you should use vcov instead.

cluster
Tells how to cluster the standard-errors (if clustering is requested). Can be either a list of vectors, a character vector of variable names, a formula or an integer vector. Assume we want to perform 2-way clustering over var1 and var2 contained in the data.frame base used for the estimation. All the following cluster arguments are valid and do the same thing: cluster = base[, c("var1", "var2")], cluster = c("var1", "var2"), cluster = ~var1+var2. If the two variables were used as fixed-effects in the estimation, you can leave it blank with vcov = "twoway" (assuming var1 [resp. var2] was the 1st [res. 2nd] fixed-effect). You can interact two variables using ^ with the following syntax: cluster = ~var1^var2 or cluster = "var1^var2".

ssc
An object of class ssc.type obtained with the function ssc. Represents how the degree of freedom correction should be done.You must use the function ssc for this argument. The arguments and defaults of the function ssc are: adj = TRUE, fixef.K="nested", cluster.adj = TRUE, cluster.df = "conventional", t.df = "conventional", fixef.force_exact=FALSE). See the help of the function ssc for details.

.vcov
A user provided covariance matrix or a function computing this matrix. If a matrix, it must be a square matrix of the same number of rows as the number of variables estimated. If a function, it must return the previously mentioned matrix.

stage
Can be equal to 2 (default), 1, 1:2 or 2:1. Only used if the object is an IV estimation: defines the stage to which summary should be applied. If stage = 1 and there are multiple endogenous regressors or if stage is of length 2, then an object of class fixest_multi is returned.

lean
Logical, default is FALSE. Used to reduce the (memory) size of the summary object. If TRUE, then all objects of length N (the number of observations) are removed from the result. Note that some fixest methods may consequently not work when applied to the summary.

n
Integer, default is 1000. Number of coefficients to display when the print method is used.

...
Not currently used.

Value

It returns either an object of class fixest_multi (if type equals short or long), either a data.frame (if type equals compact or se_compact).

See also

Examples

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res = feols(y ~ csw(x1, x2, x3), base, split = ~species)

# By default, the type is "short"
# You can still use the arguments from summary.fixest
summary(res, se = "hetero")

#> Standard-errors: Heteroskedasticity-robust 
#> 
#> ### Sample: setosa
#> 
#> Expl. vars.: x1
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  2.63900   0.298624  8.8372 1.23e-11 ***
#> x1           0.69049   0.085903  8.0380 1.93e-10 ***
#> ---
#> Expl. vars.: x1 + x2
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 2.303700   0.433561  5.3135 2.89000e-06 ***
#> x1          0.667416   0.092247  7.2351 3.60000e-09 ***
#> x2          0.283419   0.264794  1.0703 2.89933e-01    
#> ---
#> Expl. vars.: x1 + x2 + x3
#>             Estimate Std. Error  t value    Pr(>|t|)    
#> (Intercept) 2.351900   0.437230 5.379100 2.44000e-06 ***
#> x1          0.654835   0.092468 7.081800 6.87000e-09 ***
#> x2          0.237560   0.275270 0.863009 3.92610e-01    
#> x3          0.252126   0.284622 0.885827 3.80318e-01    
#> 
#> ### Sample: versicolor
#> 
#> Expl. vars.: x1
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 3.539700   0.617019  5.7368 6.31e-07 ***
#> x1          0.865078   0.220701  3.9197 2.81e-04 ***
#> ---
#> Expl. vars.: x1 + x2
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 2.116400   0.431296  4.9071 1.20000e-05 ***
#> x1          0.247642   0.176500  1.4031 1.67167e-01    
#> x2          0.735587   0.111386  6.6040 3.26000e-08 ***
#> ---
#> Expl. vars.: x1 + x2 + x3
#>              Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept)  1.895500   0.434294  4.3646 0.000071000 ***
#> x1           0.386858   0.207117  1.8678 0.068165000 .  
#> x2           0.908337   0.159800  5.6842 0.000000859 ***
#> x3          -0.679224   0.436600 -1.5557 0.126631000    
#> 
#> ### Sample: virginica
#> 
#> Expl. vars.: x1
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 3.906800   0.760748  5.1355 5.07e-06 ***
#> x1          0.901534   0.246200  3.6618 6.23e-04 ***
#> ---
#> Expl. vars.: x1 + x2
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 0.624782   0.538060  1.1612 2.51434e-01    
#> x1          0.259954   0.130150  1.9973 5.15960e-02 .  
#> x2          0.934819   0.076962 12.1460 4.20000e-16 ***
#> ---
#> Expl. vars.: x1 + x2 + x3
#>              Estimate Std. Error   t value    Pr(>|t|)    
#> (Intercept)  0.699883   0.553402  1.264700 2.12352e-01    
#> x1           0.330337   0.122494  2.696800 9.75100e-03 ** 
#> x2           0.945536   0.080881 11.691000 2.26000e-15 ***
#> x3          -0.169753   0.210310 -0.807154 4.23732e-01    

summary(res, type = "long")

#> 
#> ### Sample: setosa
#> 
#> Expl. vars.: x1
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  2.63900   0.310014  8.5125 3.74e-11 ***
#> x1           0.69049   0.089899  7.6807 6.71e-10 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.233723   Adj. R2: 0.542029
#> 
#> Expl. vars.: x1 + x2
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value   Pr(>|t|)    
#> (Intercept) 2.303700   0.385294  5.9792 2.8900e-07 ***
#> x1          0.667416   0.090356  7.3865 2.1300e-09 ***
#> x2          0.283419   0.197224  1.4370 1.5733e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.228751   Adj. R2: 0.551971
#> 
#> Expl. vars.: x1 + x2 + x3
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error  t value    Pr(>|t|)    
#> (Intercept) 2.351900   0.392868 5.986500 3.03000e-07 ***
#> x1          0.654835   0.092447 7.083300 6.83000e-09 ***
#> x2          0.237560   0.208019 1.142000 2.59359e-01    
#> x3          0.252126   0.346864 0.726873 4.70987e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.227449   Adj. R2: 0.547429
#> 
#> ### Sample: versicolor
#> 
#> Expl. vars.: x1
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 3.539700   0.562874  6.2887 9.07e-08 ***
#> x1          0.865078   0.201938  4.2839 8.80e-05 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.434612   Adj. R2: 0.261511
#> 
#> Expl. vars.: x1 + x2
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 2.116400   0.494256  4.2821 0.000091000 ***
#> x1          0.247642   0.186839  1.3254 0.191435000    
#> x2          0.735587   0.124768  5.8956 0.000000387 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.329521   Adj. R2: 0.566438
#> 
#> Expl. vars.: x1 + x2 + x3
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>              Estimate Std. Error t value   Pr(>|t|)    
#> (Intercept)  1.895500   0.507055  3.7383 0.00051100 ***
#> x1           0.386858   0.204545  1.8913 0.06489000 .  
#> x2           0.908337   0.165432  5.4907 0.00000167 ***
#> x3          -0.679224   0.435382 -1.5601 0.12559900    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.321135   Adj. R2: 0.579273
#> 
#> ### Sample: virginica
#> 
#> Expl. vars.: x1
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 3.906800   0.757061  5.1605 4.66e-06 ***
#> x1          0.901534   0.253106  3.5619 8.43e-04 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.559836   Adj. R2: 0.192579
#> 
#> Expl. vars.: x1 + x2
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 0.624782   0.524867  1.1904 2.39882e-01    
#> x1          0.259954   0.153338  1.6953 9.66340e-02 .  
#> x2          0.934819   0.089602 10.4330 8.01000e-14 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.307439   Adj. R2: 0.75132
#> 
#> Expl. vars.: x1 + x2 + x3
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: IID 
#>              Estimate Std. Error   t value    Pr(>|t|)    
#> (Intercept)  0.699883   0.533601  1.311600 1.96156e-01    
#> x1           0.330337   0.174329  1.894900 6.44000e-02 .  
#> x2           0.945536   0.090722 10.422000 1.07000e-13 ***
#> x3          -0.169753   0.198072 -0.857023 3.95875e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.305014   Adj. R2: 0.749908

summary(res, type = "compact")

#>   lhs     sample          rhs     (Intercept)                x1
#> 1   y setosa     x1           2.64*** (0.31)   0.69*** (0.0899)
#> 2   y setosa     x1 + x2       2.3*** (0.385) 0.667*** (0.0904)
#> 3   y setosa     x1 + x2 + x3 2.35*** (0.393) 0.655*** (0.0924)
#> 4   y versicolor x1           3.54*** (0.563) 0.865*** (0.202) 
#> 5   y versicolor x1 + x2      2.12*** (0.494)    0.248 (0.187) 
#> 6   y versicolor x1 + x2 + x3  1.9*** (0.507)   0.387. (0.205) 
#> 7   y virginica  x1           3.91*** (0.757) 0.902*** (0.253) 
#> 8   y virginica  x1 + x2        0.625 (0.525)    0.26. (0.153) 
#> 9   y virginica  x1 + x2 + x3     0.7 (0.534)    0.33. (0.174) 
#>                  x2             x3
#> 1                                 
#> 2    0.283 (0.197)                
#> 3    0.238 (0.208)   0.252 (0.347)
#> 4                                 
#> 5 0.736*** (0.125)                
#> 6 0.908*** (0.165)  -0.679 (0.435)
#> 7                                 
#> 8 0.935*** (0.0896)               
#> 9 0.946*** (0.0907)  -0.17 (0.198)

summary(res, type = "se_compact")

#>   lhs     sample          rhs (Intercept) (Intercept)__se        x1     x1__se
#> 1   y setosa     x1             2.6390012       0.3100143 0.6904897 0.08989888
#> 2   y setosa     x1 + x2        2.3037382       0.3852942 0.6674162 0.09035581
#> 3   y setosa     x1 + x2 + x3   2.3518898       0.3928675 0.6548350 0.09244742
#> 4   y versicolor x1             3.5397347       0.5628736 0.8650777 0.20193757
#> 5   y versicolor x1 + x2        2.1164314       0.4942556 0.2476422 0.18683892
#> 6   y versicolor x1 + x2 + x3   1.8955395       0.5070552 0.3868576 0.20454490
#> 7   y virginica  x1             3.9068365       0.7570605 0.9015345 0.25310551
#> 8   y virginica  x1 + x2        0.6247824       0.5248675 0.2599540 0.15333757
#> 9   y virginica  x1 + x2 + x3   0.6998830       0.5336009 0.3303370 0.17432873
#>          x2     x2__se         x3    x3__se
#> 1        NA         NA         NA        NA
#> 2 0.2834193 0.19722377         NA        NA
#> 3 0.2375602 0.20801921  0.2521257 0.3468636
#> 4        NA         NA         NA        NA
#> 5 0.7355868 0.12476776         NA        NA
#> 6 0.9083370 0.16543248 -0.6792238 0.4353821
#> 7        NA         NA         NA        NA
#> 8 0.9348189 0.08960197         NA        NA
#> 9 0.9455356 0.09072204 -0.1697527 0.1980724