Summary for fixest_multi objects

Summary information for fixest_multi objects. In particular, this is used to specify the type of standard-errors to be computed.

# S3 method for fixest_multi
summary(
  object,
  type = "short",
  se = NULL,
  cluster = NULL,
  dof = NULL,
  .vcov,
  stage = 2,
  lean = FALSE,
  n,
  ...
)

Arguments

object
A fixest_multi object, obtained from a fixest estimation leading to multiple results.

type
A character either equal to "short", "long", "compact", or "se_compact". If short, only the table of coefficients is displayed for each estimation. If long, then the full results are displayed for each estimation. If compact, a data.frame is returned with one line per model and the formatted coefficients + standard-errors in the columns. If se_compact, a data.frame is returned with one line per model, one numeric column for each coefficient and one numeric column for each standard-error.

se
Character scalar. Which kind of standard error should be computed: “standard”, “hetero”, “cluster”, “twoway”, “threeway” or “fourway”? By default if there are clusters in the estimation: se = "cluster", otherwise se = "standard". Note that this argument can be implicitly deduced from the argument cluster.

cluster
Tells how to cluster the standard-errors (if clustering is requested). Can be either a list of vectors, a character vector of variable names, a formula or an integer vector. Assume we want to perform 2-way clustering over var1 and var2 contained in the data.frame base used for the estimation. All the following cluster arguments are valid and do the same thing: cluster = base[, c("var1", "var2")], cluster = c("var1", "var2"), cluster = ~var1+var2. If the two variables were used as clusters in the estimation, you could further use cluster = 1:2 or leave it blank with se = "twoway" (assuming var1 [resp. var2] was the 1st [res. 2nd] cluster). You can interact two variables using ^ with the following syntax: cluster = ~var1^var2 or cluster = "var1^var2".

dof
An object of class dof.type obtained with the function dof. Represents how the degree of freedom correction should be done.You must use the function dof for this argument. The arguments and defaults of the function dof are: adj = TRUE, fixef.K="nested", cluster.adj = TRUE, cluster.df = "conventional", t.df = "conventional", fixef.force_exact=FALSE). See the help of the function dof for details.

.vcov
A user provided covariance matrix or a function computing this matrix. If a matrix, it must be a square matrix of the same number of rows as the number of variables estimated. If a function, it must return the previously mentioned matrix.

stage
Can be equal to 2 (default), 1, 1:2 or 2:1. Only used if the object is an IV estimation: defines the stage to which summary should be applied. If stage = 1 and there are multiple endogenous regressors or if stage is of length 2, then an object of class fixest_multi is returned.

lean
Logical, default is FALSE. Used to reduce the (memory) size of the summary object. If TRUE, then all objects of length N (the number of observations) are removed from the result. Note that some fixest methods may consequently not work when applied to the summary.

n
Integer, default is missing (means Inf). Number of coefficients to display when the print method is used.

...
Not currently used.

Value

It returns either an object of class fixest_multi (if type equals short or long), either a data.frame (if type equals compact or se_compact).

See also

Examples

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res = feols(y ~ csw(x1, x2, x3), base, split = ~species)

# By default, the type is "short"
# You can stil use the arguments from summary.fixest
summary(res, cluster = ~ species)

#> Standard-errors: Clustered (species) 
#> 
#> ### Sample: setosa
#> 
#> Expl. vars.: x1
#>             Estimate Std. Error t value Pr(>|t|) 
#> (Intercept)  2.63900        Inf       0        1 
#> x1           0.69049        Inf       0        1 
#> ---
#> Expl. vars.: x1 + x2
#>             Estimate Std. Error t value Pr(>|t|) 
#> (Intercept) 2.303700        Inf       0        1 
#> x1          0.667416        Inf       0        1 
#> x2          0.283419        Inf       0        1 
#> ---
#> Expl. vars.: x1 + x2 + x3
#>             Estimate Std. Error t value Pr(>|t|) 
#> (Intercept) 2.351900        Inf       0        1 
#> x1          0.654835        Inf       0        1 
#> x2          0.237560        Inf       0        1 
#> x3          0.252126        Inf       0        1 
#> 
#> ### Sample: versicolor
#> 
#> Expl. vars.: x1
#>             Estimate Std. Error t value Pr(>|t|) 
#> (Intercept) 3.539700        Inf       0        1 
#> x1          0.865078        Inf       0        1 
#> ---
#> Expl. vars.: x1 + x2
#>             Estimate Std. Error t value Pr(>|t|) 
#> (Intercept) 2.116400        Inf       0        1 
#> x1          0.247642        Inf       0        1 
#> x2          0.735587        Inf       0        1 
#> ---
#> Expl. vars.: x1 + x2 + x3
#>              Estimate Std. Error t value Pr(>|t|) 
#> (Intercept)  1.895500        Inf       0        1 
#> x1           0.386858        Inf       0        1 
#> x2           0.908337        Inf       0        1 
#> x3          -0.679224        Inf       0        1 
#> 
#> ### Sample: virginica
#> 
#> Expl. vars.: x1
#>             Estimate Std. Error t value Pr(>|t|) 
#> (Intercept) 3.906800        Inf       0        1 
#> x1          0.901534        Inf       0        1 
#> ---
#> Expl. vars.: x1 + x2
#>             Estimate Std. Error t value Pr(>|t|) 
#> (Intercept) 0.624782        Inf       0        1 
#> x1          0.259954        Inf       0        1 
#> x2          0.934819        Inf       0        1 
#> ---
#> Expl. vars.: x1 + x2 + x3
#>              Estimate Std. Error t value Pr(>|t|) 
#> (Intercept)  0.699883        Inf       0        1 
#> x1           0.330337        Inf       0        1 
#> x2           0.945536        Inf       0        1 
#> x3          -0.169753        Inf       0        1 

summary(res, type = "long")

#> 
#> ### Sample: setosa
#> 
#> Expl. vars.: x1
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  2.63900   0.298624  8.8372 1.23e-11 ***
#> x1           0.69049   0.085903  8.0380 1.93e-10 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.233723   Adj. R2: 0.542029
#> 
#> Expl. vars.: x1 + x2
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 2.303700   0.433561  5.3135 2.89000e-06 ***
#> x1          0.667416   0.092247  7.2351 3.60000e-09 ***
#> x2          0.283419   0.264794  1.0703 2.89933e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.228751   Adj. R2: 0.551971
#> 
#> Expl. vars.: x1 + x2 + x3
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>             Estimate Std. Error  t value    Pr(>|t|)    
#> (Intercept) 2.351900   0.437230 5.379100 2.44000e-06 ***
#> x1          0.654835   0.092468 7.081800 6.87000e-09 ***
#> x2          0.237560   0.275270 0.863009 3.92610e-01    
#> x3          0.252126   0.284622 0.885827 3.80318e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.227449   Adj. R2: 0.547429
#> 
#> ### Sample: versicolor
#> 
#> Expl. vars.: x1
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 3.539700   0.617019  5.7368 6.31e-07 ***
#> x1          0.865078   0.220701  3.9197 2.81e-04 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.434612   Adj. R2: 0.261511
#> 
#> Expl. vars.: x1 + x2
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 2.116400   0.431296  4.9071 1.20000e-05 ***
#> x1          0.247642   0.176500  1.4031 1.67167e-01    
#> x2          0.735587   0.111386  6.6040 3.26000e-08 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.329521   Adj. R2: 0.566438
#> 
#> Expl. vars.: x1 + x2 + x3
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>              Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept)  1.895500   0.434294  4.3646 0.000071000 ***
#> x1           0.386858   0.207117  1.8678 0.068165000 .  
#> x2           0.908337   0.159800  5.6842 0.000000859 ***
#> x3          -0.679224   0.436600 -1.5557 0.126631000    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.321135   Adj. R2: 0.579273
#> 
#> ### Sample: virginica
#> 
#> Expl. vars.: x1
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 3.906800   0.760748  5.1355 5.07e-06 ***
#> x1          0.901534   0.246200  3.6618 6.23e-04 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.559836   Adj. R2: 0.192579
#> 
#> Expl. vars.: x1 + x2
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>             Estimate Std. Error t value    Pr(>|t|)    
#> (Intercept) 0.624782   0.538060  1.1612 2.51434e-01    
#> x1          0.259954   0.130150  1.9973 5.15960e-02 .  
#> x2          0.934819   0.076962 12.1460 4.20000e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.307439   Adj. R2: 0.75132
#> 
#> Expl. vars.: x1 + x2 + x3
#> OLS estimation, Dep. Var.: y
#> Observations: 50 
#> Standard-errors: Heteroskedasticity-robust 
#>              Estimate Std. Error   t value    Pr(>|t|)    
#> (Intercept)  0.699883   0.553402  1.264700 2.12352e-01    
#> x1           0.330337   0.122494  2.696800 9.75100e-03 ** 
#> x2           0.945536   0.080881 11.691000 2.26000e-15 ***
#> x3          -0.169753   0.210310 -0.807154 4.23732e-01    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.305014   Adj. R2: 0.749908

summary(res, type = "compact")

#>   lhs     sample          rhs     (Intercept)                x1
#> 1   y setosa     x1           2.64*** (0.299)  0.69*** (0.0859)
#> 2   y setosa     x1 + x2       2.3*** (0.434) 0.667*** (0.0922)
#> 3   y setosa     x1 + x2 + x3 2.35*** (0.437) 0.655*** (0.0925)
#> 4   y versicolor x1           3.54*** (0.617) 0.865*** (0.221) 
#> 5   y versicolor x1 + x2      2.12*** (0.431)    0.248 (0.176) 
#> 6   y versicolor x1 + x2 + x3  1.9*** (0.434)   0.387. (0.207) 
#> 7   y virginica  x1           3.91*** (0.761) 0.902*** (0.246) 
#> 8   y virginica  x1 + x2        0.625 (0.538)    0.26. (0.13)  
#> 9   y virginica  x1 + x2 + x3     0.7 (0.553)   0.33** (0.122) 
#>                  x2             x3
#> 1                                 
#> 2    0.283 (0.265)                
#> 3    0.238 (0.275)   0.252 (0.285)
#> 4                                 
#> 5 0.736*** (0.111)                
#> 6 0.908*** (0.16)   -0.679 (0.437)
#> 7                                 
#> 8 0.935*** (0.077)                
#> 9 0.946*** (0.0809)  -0.17 (0.21) 

summary(res, type = "se_compact")

#>   lhs     sample          rhs (Intercept) (Intercept)__se        x1     x1__se
#> 1   y setosa     x1             2.6390012       0.2986236 0.6904897 0.08590283
#> 2   y setosa     x1 + x2        2.3037382       0.4335613 0.6674162 0.09224729
#> 3   y setosa     x1 + x2 + x3   2.3518898       0.4372300 0.6548350 0.09246765
#> 4   y versicolor x1             3.5397347       0.6170191 0.8650777 0.22070069
#> 5   y versicolor x1 + x2        2.1164314       0.4312960 0.2476422 0.17649964
#> 6   y versicolor x1 + x2 + x3   1.8955395       0.4342939 0.3868576 0.20711711
#> 7   y virginica  x1             3.9068365       0.7607485 0.9015345 0.24620049
#> 8   y virginica  x1 + x2        0.6247824       0.5380596 0.2599540 0.13015020
#> 9   y virginica  x1 + x2 + x3   0.6998830       0.5534022 0.3303370 0.12249409
#>          x2     x2__se         x3    x3__se
#> 1        NA         NA         NA        NA
#> 2 0.2834193 0.26479386         NA        NA
#> 3 0.2375602 0.27526952  0.2521257 0.2846217
#> 4        NA         NA         NA        NA
#> 5 0.7355868 0.11138587         NA        NA
#> 6 0.9083370 0.15980020 -0.6792238 0.4366002
#> 7        NA         NA         NA        NA
#> 8 0.9348189 0.07696200         NA        NA
#> 9 0.9455356 0.08088054 -0.1697527 0.2103101