Updates a fixest estimation

Updates and re-estimates a fixest model (estimated with femlm, feols or feglm). This function updates the formulas and use previous starting values to estimate a new fixest model. The data is obtained from the original call.

# S3 method for fixest
update(object, fml.update, nframes = 1, ...)

Arguments

object
A fixest object. Obtained using the functions femlm, feols or feglm.

fml.update
Changes to be made to the original argument fml. See more information on update.formula. You can add/withdraw both variables and clusters. E.g. . ~ . + x2 | . + z2 would add the variable x2 and the cluster z2 to the former estimation.

nframes
(Advanced users.) Defaults to 1. Number of frames up the stack where to perform the evaluation of the updated call. By default, this is the parent frame.

...
Other arguments to be passed to the functions femlm, feols or feglm.

Value

It returns a fixest object (see details in femlm, feols or feglm).

See also

Examples

# Example using trade data
data(trade)

# main estimation
est_pois <- femlm(Euros ~ log(dist_km) | Origin + Destination, trade)

# we add the variable log(Year)
est_2 <- update(est_pois, . ~ . + log(Year))

# we add another cluster: "Product"
est_3 <- update(est_2, . ~ . | . + Product)

# we remove the cluster "Origin" and the variable log(dist_km)
est_4 <- update(est_3, . ~ . - log(dist_km) | . - Origin)

# Quick look at the 4 estimations
esttable(est_pois, est_2, est_3, est_4)

#>                                est_pois               est_2               est_3
#> log(dist_km)        -1.5173*** (0.1131) -1.5176*** (0.1132) -1.5279*** (0.1157)
#> log(Year)                               72.3689*** (6.8997) 72.6202*** (6.9826)
#> Fixed-Effects:      ------------------- ------------------- -------------------
#> Origin                              Yes                 Yes                 Yes
#> Destination                         Yes                 Yes                 Yes
#> Product                              No                  No                 Yes
#> ___________________ ___________________ ___________________ ___________________
#> Observations                     38,325              38,325              38,325
#> S.E. type                 1-way: Origin       1-way: Origin       1-way: Origin
#> Squared Corr.                     0.378               0.384               0.612
#> Pseudo R2                        0.5895              0.5929             0.76381
#> BIC                           2.444e+12           2.424e+12           1.406e+12
#>                                   est_4
#> log(dist_km)                           
#> log(Year)           70.8171*** (5.9893)
#> Fixed-Effects:      -------------------
#> Origin                               No
#> Destination                         Yes
#> Product                             Yes
#> ___________________ ___________________
#> Observations                     38,325
#> S.E. type            1-way: Destination
#> Squared Corr.                     0.179
#> Pseudo R2                       0.35377
#> BIC                           3.848e+12